% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotYVsX.R
\name{addGuestLayer}
\alias{addGuestLayer}
\title{adds limits calculates by Guest function to ggplot object}
\usage{
addGuestLayer(
  plotObject,
  deltaGuest,
  labelGuestCriteria,
  yDisplayAsAbsolute,
  geomGuestLineAttributes
)
}
\arguments{
\item{plotObject}{object to be updated}

\item{deltaGuest}{Numeric value parameter for the Guest function.}

\item{yDisplayAsAbsolute}{A logical value if FALSE the limits are calculated for the ratio predicted/observed
if TRUE limits are calculated for observed}

\item{geomGuestLineAttributes}{list of arguments passed to geom_fun}
}
\value{
The updated \code{ggplot} object
}
\description{
adds limits calculates by Guest function to ggplot object
}
\keyword{internal}
