% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{addLabels}
\alias{addLabels}
\title{adds a labels by meta data to ggplot object}
\usage{
addLabels(plotObject, mappedData)
}
\arguments{
\item{plotObject}{A \code{ggplot} object on which to add the labels}

\item{mappedData}{MappedData object with information of mapped dimensions and units}
}
\value{
updated \code{ggplot} object
}
\description{
adds a labels by meta data to ggplot object
}
\keyword{internal}
