% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_Layer.R
\name{addLLOQLayer}
\alias{addLLOQLayer}
\title{add LLOQ Layer with lloq lines}
\usage{
addLLOQLayer(
  plotObject,
  mappedData,
  layerToCall,
  useLinetypeAsAttribute,
  geomLLOQAttributes
)
}
\arguments{
\item{plotObject}{A \code{ggplot} object on which to add the plot layer}

\item{mappedData}{object of class 'MappedData', with lloq data}

\item{layerToCall}{function ggplot2 geom layer}

\item{useLinetypeAsAttribute}{boolean, if TRUE line type is set as attribute, no legend is created}

\item{geomLLOQAttributes}{additional attributes}
}
\value{
updated plot object
}
\description{
add LLOQ Layer with lloq lines
}
