% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{addOverwriteAes}
\alias{addOverwriteAes}
\title{adds and update mapping}
\usage{
addOverwriteAes(newMaps, mapping)
}
\arguments{
\item{newMaps}{new mapping entry}

\item{mapping}{Default list of aesthetic mappings to use for plot}
}
\value{
uppdated mapping
}
\description{
adds and update mapping
}
\keyword{internal}
