% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_Layer.R
\name{addXscale}
\alias{addXscale}
\title{add X-scale}
\usage{
addXscale(plotObject, xscale, xscale.args = list())
}
\arguments{
\item{plotObject}{A \code{ggplot} object on which to add the scale}

\item{xscale}{either 'linear' then \code{ggplot2::scale_x_continuous()} or 'log' then \code{ggplot2::scale_x_log10()} is used}

\item{xscale.args}{list of arguments passed to \code{ggplot2::scale_x_continuous()} or \code{ggplot2::scale_x_log10()}}
}
\value{
The updated \code{ggplot} object
}
\description{
add X-scale
}
