% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_Layer.R
\name{addXYScale}
\alias{addXYScale}
\title{Add X and Y Scale}
\usage{
addXYScale(
  plotObject,
  xScale = NULL,
  xScaleArgs = list(),
  yScale = NULL,
  yScaleArgs = list(),
  secAxis = waiver()
)
}
\arguments{
\item{plotObject}{A \code{ggplot} object on which to add the scale.}

\item{xScale}{The x-axis scale type. Available is 'linear', 'log', 'discrete'}

\item{xScaleArgs}{A list of arguments for the x-axis scale.}

\item{yScale}{The y-axis scale type. Available is 'linear', 'log'}

\item{yScaleArgs}{A list of arguments for the y-axis scale.}

\item{secAxis}{Secondary axis arguments for scale_y functions.}
}
\value{
The updated \code{ggplot} object.
}
\description{
Add X and Y scales to a \code{ggplot} object.
}
