% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_enum.R
\name{AxisScales}
\alias{AxisScales}
\title{enumeration keys for OSPSuite.plots scaling options for axis scalings}
\usage{
AxisScales
}
\description{
enumeration keys for OSPSuite.plots scaling options for axis scalings
}
