% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_enum.R
\docType{data}
\name{BINNINGMODE}
\alias{BINNINGMODE}
\title{enumeration keys for mode of Binning}
\format{
An object of class \code{list} of length 3.
}
\usage{
BINNINGMODE
}
\description{
enumeration keys for mode of Binning
}
\keyword{datasets}
