% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_Layer.R
\name{buildWatermarkGrob}
\alias{buildWatermarkGrob}
\title{Build a \code{grid::textGrob} to be used as a watermark.}
\usage{
buildWatermarkGrob(
  label,
  x = 0.5,
  y = 0.5,
  angle = 30,
  color = "grey20",
  alpha = 0.7,
  fontsize = 12
)
}
\arguments{
\item{label}{The text of the watermark - leave NULL to use the option watermark_label}

\item{x}{Horizontal location, in 0..1 coordinates}

\item{y}{Vertical location, in 0..1  coordinates}

\item{angle}{Angle of the text}

\item{color}{Color of the test}

\item{alpha}{opacity of the text}

\item{fontsize}{size of the text}
}
\value{
a \code{grid::textGrob} with the watermark and the additional attribute 'is_watermark' set to \code{TRUE}
}
\description{
Build a \code{grid::textGrob} to be used as a watermark.
}
