% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_export.R
\name{calculatePlotDimensions}
\alias{calculatePlotDimensions}
\title{Logic to calculate height and possibly
adjust width based on plot dimensions,
aspect ratio, legend positioning, etc.}
\usage{
calculatePlotDimensions(plotObject, width)
}
\arguments{
\item{plotObject}{ggplot object}

\item{width}{\code{double} width of exported file, defaults is \code{getOption('ospsuite.plots.export.width',default = 16)}
unit \code{getOption('ospsuite.plots.export.units',  default = "cm")}}
}
\value{
a list or vector with the calculated width and height
}
\description{
Logic to calculate height and possibly
adjust width based on plot dimensions,
aspect ratio, legend positioning, etc.
}
