% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{checkIfColumnIsCategorical}
\alias{checkIfColumnIsCategorical}
\title{check if a data Column is categorical}
\usage{
checkIfColumnIsCategorical(data, mapping)
}
\arguments{
\item{data}{data.frame used for mapping}

\item{mapping}{list of aesthetic mappings}
}
\value{
A \code{Flag} is true if data column is either non numerical or a factor
}
\description{
check if a data Column is categorical
}
