% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-defaults.R
\name{colorMaps}
\alias{colorMaps}
\title{Color maps}
\usage{
colorMaps
}
\description{
List with some color maps for \code{Theme} object.

The \code{ospDefault} color map is based on colors in \code{default_igv} qualitative
color palette from \code{{ggsci}} package.
}
\seealso{
Other setDefault functions:
\code{\link{Shapes}},
\code{\link[=geom_point_osp]{geom_point_osp()}},
\code{\link[=getDefaultGeomAttributes]{getDefaultGeomAttributes()}},
\code{\link[=getDefaultOptions]{getDefaultOptions()}},
\code{\link[=getOspsuite.plots.option]{getOspsuite.plots.option()}},
\code{\link[=resetDefaultColorMapDistinct]{resetDefaultColorMapDistinct()}},
\code{\link[=resetDefaultTheme]{resetDefaultTheme()}},
\code{\link[=resetDefaults]{resetDefaults()}},
\code{\link[=scale_shape_osp]{scale_shape_osp()}},
\code{\link[=scale_shape_osp_identity]{scale_shape_osp_identity()}},
\code{\link[=scale_shape_osp_manual]{scale_shape_osp_manual()}},
\code{\link[=setDefaultColorMapDistinct]{setDefaultColorMapDistinct()}},
\code{\link[=setDefaultTheme]{setDefaultTheme()}},
\code{\link[=setDefaults]{setDefaults()}},
\code{\link[=setOspsuite.plots.option]{setOspsuite.plots.option()}},
\code{\link[=stat_qq_osp]{stat_qq_osp()}}
}
\concept{setDefault functions}
