% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CombinedPlot.R
\docType{class}
\name{CombinedPlot}
\alias{CombinedPlot}
\title{CombinedPlot}
\description{
This class represents a combined plot object that includes a plot and an optional table.
It provides methods to get and set the plot and table objects, as well as to print the combined output.
}
\examples{
\dontrun{

# Set watermark option first (required)
options(ospsuite.plots.watermark_enabled = TRUE)

# Create a new CombinedPlot instance
combinedPlotInstance <- CombinedPlot$new(plotObject = myPlotObject, tableObject = myTableObject)

# Print the combined plot and table
print(combinedPlotInstance)
# or simply
combinedPlotInstance
}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{plotObject}}{A ggplot object representing the main plot.}

\item{\code{tableObject}}{A ggplot object representing the table.}

\item{\code{relWidths}}{A numeric vector of length 2 specifying the relative widths of the plot and table.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CombinedPlot-new}{\code{CombinedPlot$new()}}
\item \href{#method-CombinedPlot-combined}{\code{CombinedPlot$combined()}}
\item \href{#method-CombinedPlot-print}{\code{CombinedPlot$print()}}
\item \href{#method-CombinedPlot-clone}{\code{CombinedPlot$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CombinedPlot-new"></a>}}
\if{latex}{\out{\hypertarget{method-CombinedPlot-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CombinedPlot$new(plotObject = ggplot(), tableObject = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{plotObject}}{A ggplot object for the main plot.
Combine the combined plot and table

This method combines the plot and table into a single output and displays it.}

\item{\code{tableObject}}{A ggplot object for the table.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CombinedPlot-combined"></a>}}
\if{latex}{\out{\hypertarget{method-CombinedPlot-combined}{}}}
\subsection{Method \code{combined()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CombinedPlot$combined()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A ggplot object representing the combined plot and table
Print the combined plot and table

This method overrides the default print function to display the combined output.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CombinedPlot-print"></a>}}
\if{latex}{\out{\hypertarget{method-CombinedPlot-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CombinedPlot$print()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Invisibly returns the combined ggplot object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CombinedPlot-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CombinedPlot-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CombinedPlot$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
