% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{constructLabelWithUnit}
\alias{constructLabelWithUnit}
\title{Construct a Label with Unit}
\usage{
constructLabelWithUnit(label, unit)
}
\arguments{
\item{label}{A character string representing the label. It should not be NULL.}

\item{unit}{A character string representing the unit. It can be NULL or an empty string.}
}
\value{
A character string that combines the label and the unit, formatted as
"label \link{unit}", or just the label if the unit is empty or NULL.
}
\description{
This function constructs a label by appending a unit in square brackets
if both the label and unit are provided. If the unit is empty or NULL,
only the label is returned.
}
\examples{
constructLabelWithUnit("Temperature", "Celsius") # Returns "Temperature [Celsius]"
constructLabelWithUnit("Length", "") # Returns "Length"
constructLabelWithUnit(NULL, "kg") # Returns NULL

}
