% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotYVsX.R
\name{countEntriesInBetween}
\alias{countEntriesInBetween}
\title{Count Entries Between Specified Limits}
\usage{
countEntriesInBetween(
  yColumn,
  xColumn,
  comparisonLineVector,
  deltaGuest,
  addGuestLimits,
  yDisplayAsAbsolute
)
}
\arguments{
\item{yColumn}{A numeric vector containing the Y values to count.}

\item{xColumn}{A numeric vector containing the X values to count.}

\item{comparisonLineVector}{A list of numeric values defining the comparison limits.}

\item{deltaGuest}{A numeric value parameter for the Guest function.}

\item{addGuestLimits}{A boolean indicating whether to include guest limits in the counting.}

\item{yDisplayAsAbsolute}{A boolean indicating whether to consider absolute values for Y.}
}
\value{
A list containing counts of entries that fall within the specified limits, including guest criteria if applicable.
}
\description{
This function counts the number of entries within specified limits for the given X and Y columns.
It calculates the counts based on the comparison line vector and guest limits, if applicable.
}
\keyword{internal}
