% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotYVsX.R
\name{createDDIInteractionPlot}
\alias{createDDIInteractionPlot}
\title{Create Drug-Drug Interaction (DDI) Interaction Plot}
\usage{
createDDIInteractionPlot(
  data = NULL,
  mapping = NULL,
  xyscale = AxisScales$log,
  comparisonLineVector = getFoldDistanceList(c(2)),
  deltaGuest = 1,
  asSquarePlot = TRUE,
  ...
)
}
\arguments{
\item{data}{A data frame containing the data to be plotted, which should include
the variables referenced in the mapping. The data frame must contain
at least one row of data for the plot to be generated.}

\item{mapping}{A mapping of aesthetics to variables in the data. This should be a list
containing ggplot2 aesthetic mappings such as \code{x} and \code{y}, linking them to
the respective columns in \code{plotData}.}

\item{xyscale}{Either "linear" or "log" scale for the X and Y axes.}

\item{comparisonLineVector}{A vector defining the comparison lines.}

\item{deltaGuest}{A numeric value parameter for the Guest function.}

\item{asSquarePlot}{A boolean; if true, the plot is returned as a square plot with aspect ratio = 1 and fixed ratios.}

\item{...}{Additional arguments that can be passed to the underlying plotting functions.}
}
\value{
A \code{ggplot} object representing the DDI interaction plot, which can be further
customized or saved using standard ggplot2 functions.
}
\description{
This function generates a ggplot object to visualize drug-drug interaction (DDI) data.
It plots predicted versus observed values on a log-log scale and helps interpret the data by adding
goodness of prediction lines, including the line of identity, guest criteria,
and two-fold lines to the plot.
}
