% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotForest.R
\name{createTableData}
\alias{createTableData}
\title{Combine Plot and Table}
\usage{
createTableData(plotData, tableColumns, tableLabels)
}
\arguments{
\item{plotObject}{A ggplot object representing the main plot.}

\item{tableObject}{A ggplot object representing the table.}

\item{relWidths}{A numeric vector of length 2 specifying the relative widths of the plot and table.}
}
\value{
A combined ggplot object containing both the plot and the table.
}
\description{
This function combines the plot and the table into a single output.
}
\keyword{internal}
