% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotForest.R
\name{createTableObject}
\alias{createTableObject}
\title{Create the Table Object}
\usage{
createTableObject(
  tableData,
  mapping,
  digitsToRound,
  digitsToShow,
  yFacetColumns
)
}
\arguments{
\item{tableData}{A data.table containing the data to be plotted in the table.}

\item{mapping}{A ggplot mapping object.}

\item{digitsToRound}{An integer specifying the number of digits to round in the table.}

\item{digitsToShow}{An integer specifying the number of digits to display in the table.}

\item{yFacetColumns}{A character vector of column names used for faceting on the y-axis.}
}
\value{
A ggplot object representing the table.
}
\description{
This function generates the table plot object.
}
\keyword{internal}
