% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTimeProfile.R
\name{.addCommonLimitsAndYscaleArguments}
\alias{.addCommonLimitsAndYscaleArguments}
\title{set the common axis limits to \code{MappedData} object and update y2-scale arguments}
\usage{
.addCommonLimitsAndYscaleArguments(
  simMappedData,
  obsMappedData,
  commonLimits,
  requireDualAxis,
  yscale.args,
  y2scale.args
)
}
\arguments{
\item{simMappedData}{object of class \code{MappedDataTimeprofile} for simulated data}

\item{obsMappedData}{object of class \code{MappedDataTimeprofile} for observed data}

\item{commonLimits}{common limits for simulated and observed data}

\item{requireDualAxis}{boolean if TRUE secondary axis is needed}

\item{yscale.args}{list with y2scale arguments}

\item{y2scale.args}{list with arguments for secondary y-axis}
}
\value{
list with
\code{simMappedData} adjusted object of class \code{MappedDataTimeprofile} for simulated data
\code{obsMappedData} adjusted object of class \code{MappedDataTimeprofile} for observed data
\code{yscale.args} adjusted \code{yscale.args} with common limits for primary and secondary y-axis
\code{secAxis}  secondary axis object
}
\description{
set the common axis limits to \code{MappedData} object and update y2-scale arguments
}
\keyword{internal}
