% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{.addLabels}
\alias{.addLabels}
\title{adds a labels by meta data to ggplot object}
\usage{
.addLabels(plotObject, metaData, mapping)
}
\arguments{
\item{plotObject}{A \code{ggplot} object on which to add the labels}

\item{metaData}{A named list of information about the \code{data} such as the \code{dimension} and \code{unit} of its variables.}

\item{mapping}{Default list of aesthetic mappings to use for plot}
}
\value{
updated \code{ggplot} object
}
\description{
adds a labels by meta data to ggplot object
}
\keyword{internal}
