% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTimeProfile.R
\name{.addLayersForObserveddData}
\alias{.addLayersForObserveddData}
\title{set line and ribbon layer fro simulated data}
\usage{
.addLayersForObserveddData(
  plotObject,
  obsMappedData,
  geomErrorbarAttributes,
  geomPointAttributes,
  geomLLOQAttributes,
  useLLOQLinetypeAsAttribute,
  mapSimulatedAndObserved,
  groupAesthetics
)
}
\arguments{
\item{plotObject}{An optional \code{ggplot} object on which to add the plot layers}

\item{obsMappedData}{object of class \code{MappedDataTimeprofile} for observed data}

\item{geomErrorbarAttributes}{A \code{list} with arguments which are passed on to the call \code{ggplot2::geom_errorbar}}

\item{geomPointAttributes}{A \code{list} with arguments which are passed on to the call \code{ggplot2::geom_point}}

\item{geomLLOQAttributes}{A \code{list} with arguments which are passed on to the call \code{ggplot2::geom_hline}}

\item{useLLOQLinetypeAsAttribute}{boolean if TRUE linetyp for LLOQ is set as attribute}

\item{mapSimulatedAndObserved}{table with columns observed and simulated which maps simulated and observed data
use of \code{mapSimulatedAndObserved} triggers reset of aesthetic scales after simulation layers}

\item{groupAesthetics}{vector of aesthetics, which are used for columns mapped with \code{groupby},}
}
\description{
set line and ribbon layer fro simulated data
}
\keyword{internal}
