% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTimeProfile.R
\name{.addLayersForSimulatedData}
\alias{.addLayersForSimulatedData}
\title{set line and ribbon layer fro simulated data}
\usage{
.addLayersForSimulatedData(
  plotObject,
  simMappedData,
  geomRibbonAttributes,
  geomLineAttributes,
  mapSimulatedAndObserved,
  groupAesthetics
)
}
\arguments{
\item{plotObject}{An optional \code{ggplot} object on which to add the plot layers}

\item{simMappedData}{object of class \code{MappedDataTimeprofile} for simulated data}

\item{geomRibbonAttributes}{A \code{list} with arguments which are passed on to the call \code{ggplot2::geom_ribbon}}

\item{geomLineAttributes}{A \code{list} with arguments which are passed on to the call \code{ggplot2::geom_line}}

\item{mapSimulatedAndObserved}{table with columns observed and simulated which maps simulated and observed data
use of \code{mapSimulatedAndObserved} triggers reset of aesthetic scales after simulation layers}

\item{groupAesthetics}{vector of aesthetics, which are used for columns mapped with \code{groupby},}
}
\value{
plot object wit newly added layers
}
\description{
set line and ribbon layer fro simulated data
}
\keyword{internal}
