% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTimeProfile.R
\name{.adjustLimits}
\alias{.adjustLimits}
\title{compares old and new limits}
\usage{
.adjustLimits(commonLimits_old, commonLimits_new)
}
\arguments{
\item{commonLimits_old}{list of old limits}

\item{commonLimits_new}{list of new limits}
}
\value{
updated limits
}
\description{
compares old and new limits
}
\keyword{internal}
