% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_shapes.R
\name{.asPlotShape}
\alias{.asPlotShape}
\title{converts shape name to value}
\usage{
.asPlotShape(shapes)
}
\arguments{
\item{shapes}{shape may be valid shape or name entry of list Shapes}
}
\value{
shape value
}
\description{
converts shape name to value
}
\keyword{internal}
