% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_shapes.R
\name{.buildOspShapeSpec}
\alias{.buildOspShapeSpec}
\title{Build shape specs with computed area-based scaling}
\usage{
.buildOspShapeSpec()
}
\value{
list of shape specifications with scale factors
}
\description{
Build shape specs with computed area-based scaling
}
\keyword{internal}
