% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTimeProfile.R
\name{.getMappedDataForTimeProfiles}
\alias{.getMappedDataForTimeProfiles}
\title{prepares mapped Data object for plotting}
\usage{
.getMappedDataForTimeProfiles(
  data,
  mapping,
  observedData,
  observedMapping,
  metaData,
  xscale,
  yscale,
  yscale.args,
  y2scale,
  y2scale.args,
  groupAesthetics,
  mapSimulatedAndObserved
)
}
\arguments{
\item{data}{data.frame with simulated data will be displayed as lines with ribbons}

\item{mapping}{a list of aesthetic mappings to use for plot,
additional to \code{{ggplot2}} aesthetics, the aesthetics \code{groupby},\code{error},\code{error_relative},\code{lloq}, \code{mdv}, \code{y2axis} are available,
see vignettes for more details and examples}

\item{observedData}{data.frame with observed data  will be displayed as points with error-bars}

\item{observedMapping}{a list of aesthetic mappings to use for observed data,
per default is is set to mapping. So if both data sets have the same mapping, use only \code{mapping},
if a different mapping is necessary use \code{mapping} and \code{observedMapping}}

\item{metaData}{A named list of information about \code{data} such as the \code{dimension} and \code{unit} of its variables.}

\item{xscale}{either 'linear' then \code{ggplot2::scale_x_continuous()} or 'log' then \code{ggplot2::scale_x_log10()} is used}

\item{yscale}{either 'linear' then \code{ggplot2::scale_y_continuous()} or 'log' then \code{ggplot2::scale_y_log10()} is used}

\item{yscale.args}{list of arguments passed to \code{ggplot2::scale_y_continuous()} or \code{ggplot2::scale_y_log10()}}

\item{y2scale}{either 'linear' the secondary axis is displayed linear, or 'log' secondary axis is displayed with log scale}

\item{y2scale.args}{list of arguments passed to \code{ggplot2::sec_axis()}, trans, break are set by code}

\item{groupAesthetics}{vector of aesthetics, which are used for columns mapped with \code{groupby},}

\item{mapSimulatedAndObserved}{table with columns observed and simulated which maps simulated and observed data
use of \code{mapSimulatedAndObserved} triggers reset of aesthetic scales after simulation layers}
}
\value{
list with entries \code{simMappedData} and \code{obsMappedData}
}
\description{
prepares mapped Data object for plotting
}
\keyword{internal}
