% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotWithWatermark.R
\name{.getWatermarkEnabled}
\alias{.getWatermarkEnabled}
\title{Helper function to get watermark enabled status.}
\usage{
.getWatermarkEnabled()
}
\description{
Helper function to get watermark enabled status.
}
\keyword{internal}
