% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-defaults.R
\name{.ospColorPalette}
\alias{.ospColorPalette}
\title{OSP color palette function}
\usage{
.ospColorPalette(n)
}
\arguments{
\item{n}{number of colors needed}
}
\value{
character vector of colors of length \code{n}
}
\description{
Reproduces the stacked-palette selection of the previous global
\verb{ggplot2.discrete.*} options: \code{colorMaps$default} for up to 6 groups,
otherwise \code{colorMaps$ospDefault}.
}
\keyword{internal}
