% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_shapes.R
\name{.ospGrob}
\alias{.ospGrob}
\title{Build grid grob for OSP shape}
\usage{
.ospGrob(name, cx, cy, half, fill, colour, stroke, alpha)
}
\arguments{
\item{name}{shape name}

\item{cx}{center x (grid unit)}

\item{cy}{center y (grid unit)}

\item{half}{half-size (grid unit)}

\item{fill}{fill color}

\item{colour}{stroke color}

\item{stroke}{stroke width}

\item{alpha}{alpha value}
}
\value{
grid grob
}
\description{
Build grid grob for OSP shape
}
\keyword{internal}
