% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_shapes.R
\name{.ospShapePalette}
\alias{.ospShapePalette}
\title{OSP shape palette function}
\usage{
.ospShapePalette(n)
}
\arguments{
\item{n}{Number of shapes needed}
}
\value{
Character vector of shape names
}
\description{
OSP shape palette function
}
\keyword{internal}
