% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_shapes.R
\docType{data}
\name{.ospShapeSpec}
\alias{.ospShapeSpec}
\title{Shape specifications for OSP shapes}
\format{
An object of class \code{list} of length 22.
}
\usage{
.ospShapeSpec
}
\description{
Shape specifications for OSP shapes
}
\keyword{internal}
