% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_shapes.R
\name{.polyVertices}
\alias{.polyVertices}
\title{Regular n-gon vertices on unit circle}
\usage{
.polyVertices(n, angleDeg = 0)
}
\arguments{
\item{n}{number of vertices}

\item{angleDeg}{angle of first vertex in degrees}
}
\value{
list with x and y coordinate vectors
}
\description{
Regular n-gon vertices on unit circle
}
\keyword{internal}
