% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_shapes.R
\name{.shoelaceArea}
\alias{.shoelaceArea}
\title{Polygon area using shoelace formula}
\usage{
.shoelaceArea(x, y)
}
\arguments{
\item{x}{x-coordinates of vertices}

\item{y}{y-coordinates of vertices}
}
\value{
area of polygon
}
\description{
Polygon area using shoelace formula
}
\keyword{internal}
