% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_shapes.R
\name{.starVertices}
\alias{.starVertices}
\title{Star vertices with alternating outer/inner radius}
\usage{
.starVertices(points = 5, innerRatio = 0.4)
}
\arguments{
\item{points}{number of star points}

\item{innerRatio}{ratio of inner to outer radius}
}
\value{
list with x and y coordinate vectors
}
\description{
Star vertices with alternating outer/inner radius
}
\keyword{internal}
