% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTimeProfile.R
\name{.validatePlotTimeProfileInputs}
\alias{.validatePlotTimeProfileInputs}
\title{Validates input for \code{plotTimeProfile} function}
\usage{
.validatePlotTimeProfileInputs(
  data,
  observedData,
  plotObject,
  metaData,
  xScale,
  xScaleArgs,
  yScale,
  yScaleArgs,
  y2Scale,
  y2ScaleArgs,
  geomLineAttributes,
  geomRibbonAttributes,
  geomPointAttributes,
  geomErrorbarAttributes,
  geomLLOQAttributes,
  groupAesthetics,
  mapSimulatedAndObserved
)
}
\arguments{
\item{data}{data.frame with simulated data will be displayed as lines with ribbons}

\item{observedData}{data.frame with observed data  will be displayed as points with error-bars}

\item{plotObject}{An optional \code{ggplot} object on which to add the plot layers}

\item{metaData}{A named list of information about \code{data} such as the \code{dimension} and \code{unit} of its variables.}

\item{xScale}{either 'linear' then \code{ggplot2::scale_x_continuous()} or 'log' then \code{ggplot2::scale_x_log10()} is used}

\item{xScaleArgs}{list of arguments passed to \code{ggplot2::scale_x_continuous()} or \code{ggplot2::scale_x_log10()}}

\item{yScale}{either 'linear' then \code{ggplot2::scale_y_continuous()} or 'log' then \code{ggplot2::scale_y_log10()} is used}

\item{yScaleArgs}{list of arguments passed to \code{ggplot2::scale_y_continuous()} or \code{ggplot2::scale_y_log10()}}

\item{y2Scale}{either 'linear' the secondary axis is displayed linear, or 'log' secondary axis is displayed with log scale}

\item{y2ScaleArgs}{list of arguments passed to \code{ggplot2::sec_axis()}, trans, break are set by code}

\item{geomLineAttributes}{A \code{list} with arguments which are passed on to the call \code{ggplot2::geom_line}}

\item{geomRibbonAttributes}{A \code{list} with arguments which are passed on to the call \code{ggplot2::geom_ribbon}}

\item{geomPointAttributes}{A \code{list} with arguments which are passed on to the call \code{ggplot2::geom_point}}

\item{geomErrorbarAttributes}{A \code{list} with arguments which are passed on to the call \code{ggplot2::geom_errorbar}}

\item{geomLLOQAttributes}{A \code{list} with arguments which are passed on to the call \code{ggplot2::geom_hline}}

\item{groupAesthetics}{vector of aesthetics, which are used for columns mapped with \code{groupby},}

\item{mapSimulatedAndObserved}{table with columns observed and simulated which maps simulated and observed data
use of \code{mapSimulatedAndObserved} triggers reset of aesthetic scales after simulation layers}
}
\description{
Validates input for \code{plotTimeProfile} function
}
\keyword{internal}
