% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exampleDataCovariates}
\alias{exampleDataCovariates}
\title{data frame which includes example data used in the vignette
\code{vignette("Histogram Plots", package = "ospsuite.plots")}}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 410 rows and 11 columns.
}
\usage{
exampleDataCovariates
}
\description{
data frame which includes example data used in the vignette
\code{vignette("Histogram Plots", package = "ospsuite.plots")}
}
\keyword{datasets}
