% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_export.R
\name{exportPlot}
\alias{exportPlot}
\title{export plot object}
\usage{
exportPlot(
  plotObject,
  filepath,
  filename,
  width = getOspsuite.plots.option(optionKey = OptionKeys$export.width),
  height = NULL,
  ...
)
}
\arguments{
\item{plotObject}{ggplot object}

\item{filepath}{name of export directory}

\item{filename}{name of file to save.
extension is added or switched to default device
\code{getOspsuite.plots.option(optionKey = OptionKeys$export.device)}}

\item{width}{\code{double} width of exported file, defaults is \code{getOption('ospsuite.plots.export.width',default = 16)}
unit \code{getOption('ospsuite.plots.export.units',  default = "cm")}}

\item{height}{\code{double} height of exported plot, if NULL the aspect ratio of plot is used,
or if no aspect ratio is defined, a square plot is exported}

\item{...}{Other arguments passed on to the ggsave}
}
\description{
saves plot object to file

For more details and examples see the vignettes:
\itemize{
\item \code{vignette("ospsuite.plots", package = "ospsuite.plots")}
}
}
