% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-errorbar-osp.R
\name{geom_errorbar_osp}
\alias{geom_errorbar_osp}
\title{OSP Errorbar Layer}
\usage{
geom_errorbar_osp(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  ...,
  orientation = NA,
  width = 2,
  lineend = "butt",
  na.rm = FALSE,
  show.legend = FALSE,
  inherit.aes = TRUE
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{aes()}}. If specified and
\code{inherit.aes = TRUE} (the default), it is combined with the default mapping
at the top level of the plot. You must supply \code{mapping} if there is no plot
mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[ggplot2:ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[ggplot2:fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame}, and
will be used as the layer data. A \code{function} can be created
from a \code{formula} (e.g. \code{~ head(.x, 10)}).}

\item{stat}{The statistical transformation to use on the data for this layer.
When using a \verb{geom_*()} function to construct a layer, the \code{stat}
argument can be used to override the default coupling between geoms and
stats. The \code{stat} argument accepts the following:
\itemize{
\item A \code{Stat} ggproto subclass, for example \code{StatCount}.
\item A string naming the stat. To give the stat as a string, strip the
function name of the \code{stat_} prefix. For example, to use \code{stat_count()},
give the stat as \code{"count"}.
\item For more information and other ways to specify the stat, see the
\link[ggplot2:layer_stats]{layer stat} documentation.
}}

\item{position}{A position adjustment to use on the data for this layer. This
can be used in various ways, including to prevent overplotting and
improving the display. The \code{position} argument accepts the following:
\itemize{
\item The result of calling a position function, such as \code{position_jitter()}.
This method allows for passing extra arguments to the position.
\item A string naming the position adjustment. To give the position as a
string, strip the function name of the \code{position_} prefix. For example,
to use \code{position_jitter()}, give the position as \code{"jitter"}.
\item For more information and other ways to specify the position, see the
\link[ggplot2:layer_positions]{layer position} documentation.
}}

\item{...}{Other arguments passed on to \code{\link[ggplot2:layer]{layer()}}'s \code{params} argument. These
arguments broadly fall into one of 4 categories below. Notably, further
arguments to the \code{position} argument, or aesthetics that are required
can \emph{not} be passed through \code{...}. Unknown arguments that are not part
of the 4 categories below are ignored.
\itemize{
\item Static aesthetics that are not mapped to a scale, but are at a fixed
value and apply to the layer as a whole. For example, \code{colour = "red"}
or \code{linewidth = 3}. The geom's documentation has an \strong{Aesthetics}
section that lists the available options. The 'required' aesthetics
cannot be passed on to the \code{params}. Please note that while passing
unmapped aesthetics as vectors is technically possible, the order and
required length is not guaranteed to be parallel to the input data.
\item When constructing a layer using
a \verb{stat_*()} function, the \code{...} argument can be used to pass on
parameters to the \code{geom} part of the layer. An example of this is
\code{stat_density(geom = "area", outline.type = "both")}. The geom's
documentation lists which parameters it can accept.
\item Inversely, when constructing a layer using a
\verb{geom_*()} function, the \code{...} argument can be used to pass on parameters
to the \code{stat} part of the layer. An example of this is
\code{geom_area(stat = "density", adjust = 0.5)}. The stat's documentation
lists which parameters it can accept.
\item The \code{key_glyph} argument of \code{\link[ggplot2:layer]{layer()}} may also be passed on through
\code{...}. This can be one of the functions described as
\link[ggplot2:draw_key]{key glyphs}, to change the display of the layer in the legend.
}}

\item{orientation}{Orientation of the layer. \code{"x"} produces horizontal
error bars (range along the x-axis). Any other value, including \code{NA}
(default) and \code{"y"}, produces vertical error bars (range along the
y-axis).}

\item{width}{Width of the error bar caps in mm units. Default: \code{2}.}

\item{lineend}{Line end style (round, butt, square).}

\item{na.rm}{Missing values in the range aesthetics are dropped (the
affected cap is simply not drawn) for both \code{TRUE} and \code{FALSE}. No warning
is emitted in either case. The argument is kept for consistency with the
\code{\link[ggplot2:geom_errorbar]{ggplot2::geom_errorbar()}} interface.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display. To include legend keys for all levels, even
when no data exists, use \code{TRUE}.  If \code{NA}, all levels are shown in legend,
but unobserved levels are omitted.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[ggplot2:annotation_borders]{annotation_borders()}}.}
}
\value{
A ggplot2 layer that can be added to a plot.
}
\description{
A geom that renders error bars with cap \code{width} specified in \strong{mm} units,
keeping it visually consistent with the \code{linewidth} aesthetic, which is
also expressed in \strong{mm}. Unlike \code{\link[ggplot2:geom_errorbar]{ggplot2::geom_errorbar()}},
the cap width is independent of the data coordinate range or axis scale.

Vertical orientation (\code{aes(x, ymin, ymax)}) is used by default.
Pass \code{orientation = "x"} for horizontal error bars (\code{aes(y, xmin, xmax)}).
}
\examples{
library(ggplot2)
df <- data.frame(
  x    = 1:3,
  y    = c(1, 2, 3),
  ymin = c(0.5, 1.5, 2.5),
  ymax = c(1.5, 2.5, 3.5)
)
ggplot(df, aes(x, y, ymin = ymin, ymax = ymax)) +
  geom_errorbar_osp(width = 2, linewidth = 0.8)
}
\seealso{
Other layers:
\code{\link{GeomErrorbarOsp}},
\code{\link{GeomPointOsp}},
\code{\link[=geom_point_osp]{geom_point_osp()}},
\code{\link[=stat_qq_osp]{stat_qq_osp()}}
}
\concept{layers}
