% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_shapes.R
\name{geomPointUnicode}
\alias{geomPointUnicode}
\title{geomTLFPoint}
\usage{
geomPointUnicode(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{mapping}{mapping from \code{ggplot2} package as provided by \code{aes()}}

\item{data}{data.frame}

\item{stat}{stat name from \code{ggplot2}}

\item{position}{position name from \code{ggplot2}}

\item{na.rm}{a logical value indicating}

\item{show.legend}{= NA,}

\item{inherit.aes}{a logical value indicating if aesthetics are inherited}

\item{...}{other arguments.}
}
\description{
geom similar to \code{geom_point()} but that leverage fonts to draw its shapes
}
\seealso{
Other setDefault functions: 
\code{\link{ColorMaps}},
\code{\link{Shapes}},
\code{\link{getDefaultGeomAttributes}()},
\code{\link{getDefaultOptions}()},
\code{\link{resetDefaultColorMapDistinct}()},
\code{\link{resetDefaultShapeDiscrete}()},
\code{\link{resetDefaults}()},
\code{\link{setDefaultColorMapDistinct}()},
\code{\link{setDefaultShapeDiscrete}()},
\code{\link{setDefaultTheme}()},
\code{\link{setDefaults}()}
}
\concept{setDefault functions}
