% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{getDataForAesthetic}
\alias{getDataForAesthetic}
\title{fetch data column for given aesthetic}
\usage{
getDataForAesthetic(aesthetic, data, mapping, stopIfNull = FALSE)
}
\arguments{
\item{aesthetic}{Aesthetic fro which the the data is required}

\item{data}{data.frame with data}

\item{mapping}{Default list of aesthetic mappings to use for plot}

\item{stopIfNull}{If TRUE error will be thrown if data column does not exists,
IF FALSE for non existing columns NULL will be returned}
}
\value{
vector with values corresponding to the mapped data column
}
\description{
fetch data column for given aesthetic
}
\keyword{internal}
