% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-defaults.R
\name{getDefaultGeomAttributes}
\alias{getDefaultGeomAttributes}
\title{get the defaults for the geom attributes used as defaults in plot functions
see \code{vignette("ospsuite.plots", package = "ospsuite.plots")} how to change defaults}
\usage{
getDefaultGeomAttributes(geom)
}
\arguments{
\item{geom}{type of geom to return attributes}
}
\value{
list with default attributes
}
\description{
get the defaults for the geom attributes used as defaults in plot functions
see \code{vignette("ospsuite.plots", package = "ospsuite.plots")} how to change defaults
}
\seealso{
Other setDefault functions:
\code{\link{colorMaps}},
\code{\link[=getDefaultOptions]{getDefaultOptions()}},
\code{\link[=getOspsuite.plots.option]{getOspsuite.plots.option()}},
\code{\link[=resetDefaultColorMapDistinct]{resetDefaultColorMapDistinct()}},
\code{\link[=resetDefaultTheme]{resetDefaultTheme()}},
\code{\link[=resetDefaults]{resetDefaults()}},
\code{\link[=setDefaultColorMapDistinct]{setDefaultColorMapDistinct()}},
\code{\link[=setDefaultTheme]{setDefaultTheme()}},
\code{\link[=setDefaults]{setDefaults()}},
\code{\link[=setOspsuite.plots.option]{setOspsuite.plots.option()}},
\code{\link[=theme_osp]{theme_osp()}}
}
\concept{setDefault functions}
