% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotYVsX.R
\name{getGuestLimits}
\alias{getGuestLimits}
\title{Calculate Limits for DDI Ratio according to Guest et al.}
\usage{
getGuestLimits(x, deltaGuest = 1, yDisplayAsAbsolute = FALSE, asLower = TRUE)
}
\arguments{
\item{x}{A numeric vector representing the observed values.}

\item{deltaGuest}{Numeric value parameter for the Guest function.}

\item{yDisplayAsAbsolute}{A logical value if FALSE the limits are calculated for the ratio predicted/observed
if TRUE limits are calculated for observed}

\item{asLower}{A logical value indicating whether to calculate lower limits (default is TRUE).}
}
\value{
A numeric vector representing the calculated limits for the DDI ratio.
}
\description{
This function calculates the limits according to Guest et al.
for the DDI ratio based on the provided parameters.
}
\references{
\url{https://pubmed.ncbi.nlm.nih.gov/21036951}
}
