% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-defaults.R
\name{getOspsuite.plots.option}
\alias{getOspsuite.plots.option}
\title{returns an option value for a option defined by the package OSPSuite.plots}
\usage{
getOspsuite.plots.option(optionKey)
}
\arguments{
\item{optionKey}{identifier of option}
}
\value{
option value
}
\description{
returns an option value for a option defined by the package OSPSuite.plots
}
\examples{
\dontrun{
getOspsuite.plots.option(optionKey = OptionKeys$watermarkEnabled)
}
}
\seealso{
Other setDefault functions:
\code{\link{colorMaps}},
\code{\link[=getDefaultGeomAttributes]{getDefaultGeomAttributes()}},
\code{\link[=getDefaultOptions]{getDefaultOptions()}},
\code{\link[=resetDefaultColorMapDistinct]{resetDefaultColorMapDistinct()}},
\code{\link[=resetDefaultTheme]{resetDefaultTheme()}},
\code{\link[=resetDefaults]{resetDefaults()}},
\code{\link[=setDefaultColorMapDistinct]{setDefaultColorMapDistinct()}},
\code{\link[=setDefaultTheme]{setDefaultTheme()}},
\code{\link[=setDefaults]{setDefaults()}},
\code{\link[=setOspsuite.plots.option]{setOspsuite.plots.option()}},
\code{\link[=theme_osp]{theme_osp()}}
}
\concept{setDefault functions}
