% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_export.R
\name{getPlotDimensions}
\alias{getPlotDimensions}
\title{Get dimensions of plot components}
\usage{
getPlotDimensions(plotObject, exportunits, nCol, nRow, nPanel)
}
\arguments{
\item{plotObject}{A ggplot object.}

\item{exportunits}{Units of the exported figure.}

\item{nCol}{Number of columns for panel plots.}

\item{nRow}{Number of rows for panel plots.}

\item{nPanel}{Number of panels for panel plots.}
}
\value{
A list with dimensions of plot components.
}
\description{
Get dimensions of plot components
}
\keyword{internal}
