% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotWithWatermark.R
\name{ggplotWithWatermark}
\alias{ggplotWithWatermark}
\title{Create a ggplot with an optional watermark}
\usage{
ggplotWithWatermark(...)
}
\arguments{
\item{...}{Arguments to be passed to \code{ggplot()}, such as data and aesthetics. This allows for flexibility in creating different types of plots.}
}
\value{
A ggplot object, which may have the class "ggWatermark" if the watermark is enabled. The object can be printed or further modified as needed.
}
\description{
This function creates a ggplot object and adds a watermark if the watermark option is enabled.
The watermark can be customized with various options such as position, angle, font size, color, and transparency.
}
\details{
If the watermark feature is enabled, the resulting ggplot object will include a watermark overlay when printed.
The watermark's properties are determined by options set in the Ospsuite plotting configuration.
}
\examples{
# Example usage with watermark enabled
plot_with_watermark <- ggplotWithWatermark(data = mtcars, aes(x = wt, y = mpg)) +
  geom_point()
print(plot_with_watermark)


}
