% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{labelsForPercentile}
\alias{labelsForPercentile}
\title{creates labels for percentiles}
\usage{
labelsForPercentile(percentiles, suffix = " percentile")
}
\arguments{
\item{percentiles}{numeric input}

\item{suffix}{common ending}
}
\value{
character vector with labels
}
\description{
convert a numeric percentile to a label
e.g. c(25,75)   to c('25th percentile', '75th percentile')
}
