% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_Layer.R
\name{layer_watermark}
\alias{layer_watermark}
\title{Create a watermark layer for a ggplot object.}
\usage{
layer_watermark(
  label = NULL,
  x = NULL,
  y = NULL,
  angle = NULL,
  color = NULL,
  alpha = NULL,
  fontsize = NULL,
  show = NULL
)
}
\arguments{
\item{label}{Passed to \code{buildWatermarkGrob}.}

\item{x}{Passed to \code{buildWatermarkGrob}.}

\item{y}{Passed to \code{buildWatermarkGrob}.}

\item{angle}{Passed to \code{buildWatermarkGrob}.
TRUE/FALSE for explicit control, otherwise use option watermark_enabled}

\item{color}{Passed to \code{buildWatermarkGrob}.}

\item{alpha}{Passed to \code{buildWatermarkGrob}.}

\item{fontsize}{Passed to \code{buildWatermarkGrob}.}

\item{show}{Determines whether the watermark will be created. Set to}
}
\value{
a ggplot2 layer with the watermark (if show is \code{TRUE} or (\code{NULL} and
option 'ospsuite.plots.watermark_enabled' is \code{TRUE})), otherwise an empty layer
}
\description{
Create a watermark layer for a ggplot object.
}
