% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MappedDataBoxplot.R
\name{MappedDataBoxplot}
\alias{MappedDataBoxplot}
\title{MappedDataBoxplot}
\description{
R6 class for mapping variable to \code{data}
}
\seealso{
Other MappedData classes: 
\code{\link{MappedDataTimeProfile}},
\code{\link{MappedData}}
}
\concept{MappedData classes}
\section{Super class}{
\code{\link[ospsuite.plots:MappedData]{ospsuite.plots::MappedData}} -> \code{MappedDataBoxplot}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{scale}}{of x axis}

\item{\code{arguments}}{for scale of x axis}

\item{\code{Flag}}{indicates if x is mapped}

\item{\code{mapping}}{for box whisker plot}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{scale}}{of x axis}

\item{\code{arguments}}{for scale of x axis}

\item{\code{Flag}}{indicates if x is mapped}

\item{\code{mapping}}{for box whisker plot}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MappedDataBoxplot-new}{\code{MappedDataBoxplot$new()}}
\item \href{#method-MappedDataBoxplot-clone}{\code{MappedDataBoxplot$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="ospsuite.plots" data-topic="MappedData" data-id="getAestheticsForGeom"><a href='../../ospsuite.plots/html/MappedData.html#method-MappedData-getAestheticsForGeom'><code>ospsuite.plots::MappedData$getAestheticsForGeom()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MappedDataBoxplot-new"></a>}}
\if{latex}{\out{\hypertarget{method-MappedDataBoxplot-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{MappedDataBoxplot} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MappedDataBoxplot$new(
  data,
  mapping,
  groupAesthetics = NULL,
  direction = "y",
  scaleOfDirection = "linear",
  isObserved = TRUE,
  xscale,
  xscale.args
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{data.frame used for mapping}

\item{\code{mapping}}{list of aesthetic mappings}

\item{\code{groupAesthetics}}{vector of aesthetics, which are used for columns mapped with group,
use of group aesthetics triggers second axis after simulation layers}

\item{\code{direction}}{direction of plot either "x" or "y"}

\item{\code{scaleOfDirection}}{scale of direction, either "linear" or "log"}

\item{\code{isObserved}}{Flag if TRUE mappings mdv, lloq, error and error_relative are evaluated}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
MappedDataBoxplot class object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MappedDataBoxplot-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MappedDataBoxplot-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MappedDataBoxplot$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
