% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messages.R
\docType{data}
\name{messages}
\alias{messages}
\title{List of functions and strings used to signal error messages and warnings}
\format{
An object of class \code{list} of length 21.
}
\usage{
messages
}
\description{
List of functions and strings used to signal error messages and warnings
}
\keyword{internal}
