% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{metaData2DataFrame}
\alias{metaData2DataFrame}
\title{converts metaData List to a data frame
row names specify properties}
\usage{
metaData2DataFrame(metaData)
}
\arguments{
\item{metaData}{A named list of information about the \code{data} such as the \code{dimension} and \code{unit} of its variables.}
}
\value{
metaData as \code{data.frame}
}
\description{
converts metaData List to a data frame
row names specify properties
}
