% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_shapes.R
\name{ospShapeNames}
\alias{ospShapeNames}
\title{OSP Shape Names}
\usage{
ospShapeNames
}
\description{
Character vector of all available OSP shape names.
}
\seealso{
Other shapes:
\code{\link{Shapes}},
\code{\link[=scale_shape_osp]{scale_shape_osp()}},
\code{\link[=scale_shape_osp_identity]{scale_shape_osp_identity()}},
\code{\link[=scale_shape_osp_manual]{scale_shape_osp_manual()}}
}
\concept{shapes}
