% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ospsuite.plots-package.R
\docType{package}
\name{ospsuite.plots-package}
\alias{ospsuite.plots}
\alias{ospsuite.plots-package}
\title{ospsuite.plots: Library for standardized graphs}
\description{
This library provides a standardized approach to creating graphs and tables in the ospsuite context. It is based on ggplot2.
}
\author{
\strong{Maintainer}: Katrin Coboeken \email{katrin.coboeken@bayer.com}

Other contributors:
\itemize{
  \item Open-Systems-Pharmacology Community [copyright holder]
}

}
\keyword{internal}
