% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotWithWatermark.R
\name{plot.ggWatermark}
\alias{plot.ggWatermark}
\title{Create plot function for ggWatermark.}
\usage{
\method{plot}{ggWatermark}(x, ...)
}
\arguments{
\item{x}{A ggWatermark object to be printed.}

\item{...}{Additional arguments to be passed to the print method.}
}
\description{
This method allows for the ggWatermark object to be plotted using the standard print method.
}
