% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotForest.R
\name{plotForest}
\alias{plotForest}
\title{Create a Forest Plot}
\usage{
plotForest(
  plotData,
  mapping = aes(y = y, x = x, groupby = dataType),
  xLabel,
  yFacetColumns = NULL,
  xFacetColumn = NULL,
  xScale = c("linear", "log"),
  xScaleArgs = list(),
  groupAesthetics = c("color", "fill", "shape"),
  tableColumns = c("yValues", "yErrorValues"),
  tableLabels = c("M", "Variance"),
  labelWrapWidth = 10,
  digitsToRound = 2,
  digitsToShow = 2,
  withTable = is.null(xFacetColumn),
  geomPointAttributes = getDefaultGeomAttributes("Point"),
  geomErrorbarAttributes = getDefaultGeomAttributes("Errorbar"),
  facetScales = c("free_y", "free")
)
}
\arguments{
\item{plotData}{A data.table containing the data to be plotted. Must include columns specified in \code{yFacetColumns}, \code{xFacetColumn}, \code{tableColumns}, and others.}

\item{mapping}{A ggplot2 mapping object, typically created with \code{ggplot2::aes()}, to specify how variables in the data are mapped to visual properties.}

\item{xLabel}{A string representing the label for the x-axis.}

\item{yFacetColumns}{A character vector of column names used for faceting on the y-axis. Can be NULL or of length up to 2.}

\item{xFacetColumn}{A character string specifying the column name for the x-axis facet. Must be of length 1 or NULL.}

\item{xScale}{A character string indicating the scale type for the x-axis. Options are "linear" or "log".}

\item{xScaleArgs}{A list of additional arguments for customizing the x-axis scale.}

\item{groupAesthetics}{A character vector specifying aesthetics for grouping (e.g., color, fill, shape).}

\item{tableColumns}{A character vector of column names to be included in the table.}

\item{tableLabels}{A character vector of labels corresponding to \code{tableColumns}.}

\item{labelWrapWidth}{A numeric value specifying the width for label wrapping in facets.}

\item{digitsToRound}{An integer specifying the number of digits to round in the table.}

\item{digitsToShow}{An integer specifying the number of digits to display in the table.}

\item{withTable}{A logical flag indicating whether to include the table in the output. Defaults to TRUE if \code{xFacetColumn} is not NULL.}

\item{geomPointAttributes}{A list of attributes for the point geometry in the plot.}

\item{geomErrorbarAttributes}{A list of attributes for the error bar geometry in the plot.}

\item{facetScales}{A character string indicating the scales used for facets. Options are "free_y" or "free".}
}
\value{
A combined plot object containing the forest plot and the table (if applicable).
}
\description{
This function generates a forest plot with optional faceting and a corresponding table.
}
\seealso{
Other plot functions: 
\code{\link{plotBoxWhisker}()},
\code{\link{plotHistogram}()},
\code{\link{plotPredVsObs}()},
\code{\link{plotQQ}()},
\code{\link{plotRangeDistribution}()},
\code{\link{plotRatioVsCov}()},
\code{\link{plotResVsCov}()},
\code{\link{plotTimeProfile}()},
\code{\link{plotYVsX}()}
}
\concept{plot functions}
