% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHistogram.R
\name{plotHelperHistogram}
\alias{plotHelperHistogram}
\title{class to support \code{plotHistogram}}
\description{
R6 class container for functions and properties used in \code{plotHistogram}
}
\keyword{internal}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{plotAsFrequency}}{\code{Flag} to decide if data should be displayed as frequency}

\item{\code{distribution}}{\code{character} distribution to fit}

\item{\code{scaledMeanFun}}{\code{function} to display mean}

\item{\code{isStacked}}{\code{Flag} if True, frequency is displayed a per total, otherwise per group}

\item{\code{asBarPlot}}{\code{Flag} indicates plot switches to bar plot}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{getHistMapping}}{generates mapping for histogram
check if variable is a valid distribution
extract bin width and bin border of ggplot object
check if input is valid and returns function for vertical line}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-plotHelperHistogram-new}{\code{plotHelperHistogram$new()}}
\item \href{#method-plotHelperHistogram-getDistrMapping}{\code{plotHelperHistogram$getDistrMapping()}}
\item \href{#method-plotHelperHistogram-getMeanMapping}{\code{plotHelperHistogram$getMeanMapping()}}
\item \href{#method-plotHelperHistogram-clone}{\code{plotHelperHistogram$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotHelperHistogram-new"></a>}}
\if{latex}{\out{\hypertarget{method-plotHelperHistogram-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{MappedData} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{plotHelperHistogram$new(
  xscale,
  plotAsFrequency,
  asBarPlot,
  geomHistAttributes,
  distribution,
  meanFunction
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{xscale}}{scale of x -axis}

\item{\code{plotAsFrequency}}{Flag to switch display of y to frequency}

\item{\code{asBarPlot}}{Flag to switch from geom_histogram to geom_bar}

\item{\code{geomHistAttributes}}{attribute for plotting the histogram`}

\item{\code{distribution}}{distribution to fit the data}

\item{\code{meanFunction}}{function to display mean}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{plotHelperHistogram} object
generates mapping for distribution
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotHelperHistogram-getDistrMapping"></a>}}
\if{latex}{\out{\hypertarget{method-plotHelperHistogram-getDistrMapping}{}}}
\subsection{Method \code{getDistrMapping()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{plotHelperHistogram$getDistrMapping(mappedData, plotObject)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{mappedData}}{mapped data of Histogram}

\item{\code{plotObject}}{ggplot object}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
mapping
generates mapping for display of mean
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotHelperHistogram-getMeanMapping"></a>}}
\if{latex}{\out{\hypertarget{method-plotHelperHistogram-getMeanMapping}{}}}
\subsection{Method \code{getMeanMapping()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{plotHelperHistogram$getMeanMapping(mappedData)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{mappedData}}{mapped data of Histogram}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
mapping
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotHelperHistogram-clone"></a>}}
\if{latex}{\out{\hypertarget{method-plotHelperHistogram-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{plotHelperHistogram$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
