% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHistogram.R
\name{plotHistogram}
\alias{plotHistogram}
\title{Generates Histograms}
\usage{
plotHistogram(
  data,
  mapping,
  metaData = NULL,
  asBarPlot = NULL,
  geomHistAttributes = getDefaultGeomAttributes("Hist"),
  plotAsFrequency = FALSE,
  xScale = AxisScales$linear,
  xScaleArgs = list(),
  yScale = AxisScales$linear,
  yScaleArgs = list(),
  distribution = "none",
  meanFunction = "auto",
  residualScale = ResidualScales$log
)
}
\arguments{
\item{data}{data.frame with simulated data will be displayed as lines with ribbons}

\item{mapping}{a list of aesthetic mappings to use for plot,
additional to \code{{ggplot2}} aesthetics, the aesthetics \code{groupby},\code{error},\code{error_relative},\code{lloq}, \code{mdv}, \code{y2axis} are available,
see vignettes for more details and examples}

\item{metaData}{A named list of information about \code{data} such as the \code{dimension} and \code{unit} of its variables.}

\item{asBarPlot}{A \code{logical} indicating if \code{geom_histogram} should be used (for continuous data) or \code{geom_bar} (for categorical data).
If TRUE, the variables \code{distribution}, \code{meanFunction}, \code{xScale}, and \code{xScaleArgs} are ignored.}

\item{geomHistAttributes}{A \code{list} of arguments passed to \code{ggplot2::geom_histogram} (or \code{geom_bar} if \code{asBarPlot} = TRUE).}

\item{plotAsFrequency}{A \code{logical} indicating if the histogram displays frequency on the y-axis.}

\item{xScale}{either 'linear' then \code{ggplot2::scale_x_continuous()} or 'log' then \code{ggplot2::scale_x_log10()} is used}

\item{xScaleArgs}{list of arguments passed to \code{ggplot2::scale_x_continuous()} or \code{ggplot2::scale_x_log10()}}

\item{yScale}{either 'linear' then \code{ggplot2::scale_y_continuous()} or 'log' then \code{ggplot2::scale_y_log10()} is used}

\item{yScaleArgs}{list of arguments passed to \code{ggplot2::scale_y_continuous()} or \code{ggplot2::scale_y_log10()}}

\item{distribution}{Name of the distribution to fit. Available distributions are those in the \code{stats} package (see ?stats::distributions):
\code{norm}, \code{lnorm}, \code{weibull}, \code{gamma}, etc. Use \code{"none"} for no fit (default).
Shortcuts: \code{"normal"} (same as \code{"norm"}), \code{"lognormal"} (same as \code{"lnorm"}).}

\item{meanFunction}{Function selection for the display of a vertical line. Options: \code{'none'}, \code{'mean'}, \code{'geomean'}, \code{'median'}, \code{'auto'} (default).
\code{'auto'} selects \code{'mean'} for normal distribution, \code{'geomean'} for lognormal, \code{'median'} for other distributions, and \code{'none'} when no distribution fit.}

\item{residualScale}{Either "linear", "log", or "ratio" scale for residuals.}
}
\value{
A \code{ggplot} object.
}
\description{
Produces histograms with optional distribution fit.

For more details and examples see the vignettes:
\itemize{
\item \code{vignette("Histogram Plots", package = "ospsuite.plots")}
\item \code{vignette("ospsuite.plots", package = "ospsuite.plots")}
\item \code{vignette("Goodness of fit", package = "ospsuite.plots")}
}
}
\seealso{
Other plot functions: 
\code{\link{plotBoxWhisker}()},
\code{\link{plotForest}()},
\code{\link{plotPredVsObs}()},
\code{\link{plotQQ}()},
\code{\link{plotRangeDistribution}()},
\code{\link{plotRatioVsCov}()},
\code{\link{plotResVsCov}()},
\code{\link{plotTimeProfile}()},
\code{\link{plotYVsX}()}
}
\concept{plot functions}
