% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotQQ.R
\name{plotQQ}
\alias{plotQQ}
\title{generates residual quantile quantile plot}
\usage{
plotQQ(
  data,
  mapping,
  metaData = NULL,
  xscale.args = list(),
  residualScale = ResidualScales$log,
  yscale.args = list(),
  geomQQAttributes = list(),
  geomQQLineAttributes = geomQQAttributes,
  groupAesthetics = c("colour", "fill", "shape")
)
}
\arguments{
\item{data}{´data.frame` with data to plot}

\item{mapping}{a list of aesthetic mappings to use for plot,
additional to \code{{ggplot2}} aesthetics, the aesthetics \code{groupby},\code{error},\code{error_relative},\code{lloq}, \code{mdv}, \code{y2axis} are available,
see vignettes for more details and examples}

\item{metaData}{A named list of information about \code{data} such as the \code{dimension} and \code{unit} of its variables.}

\item{xscale.args}{list of arguments passed to \code{ggplot2::scale_x_continuous()} or \code{ggplot2::scale_x_log10()}}

\item{residualScale}{Either \code{"linear"} or \code{"log"} for scaling residuals.
For linear: residuals = observed - predicted. For log: residuals = log(observed) - log(predicted).
The y-axis scale remains linear in both cases.}

\item{yscale.args}{list of arguments passed to \code{ggplot2::scale_y_continuous()} or \code{ggplot2::scale_y_log10()}}

\item{geomQQAttributes}{A list of arguments passed to \code{ggplot2::stat_qq()}.}

\item{geomQQLineAttributes}{A list of arguments passed to \code{ggplot2::stat_qq_line()}.}

\item{groupAesthetics}{A character vector of aesthetic names used for grouping data points in the Q-Q plot.
Common options include \code{"colour"}, \code{"fill"}, \code{"shape"}, \code{"linetype"}, and \code{"size"}.}
}
\value{
A \code{ggplot} object
}
\description{
For details and examples see the vignettes:
\itemize{
\item \code{vignette("Goodness of fit", package = "ospsuite.plots")}
\item \code{vignette("ospsuite.plots", package = "ospsuite.plots")}
}
}
\seealso{
Other plot functions: 
\code{\link{plotBoxWhisker}()},
\code{\link{plotForest}()},
\code{\link{plotHistogram}()},
\code{\link{plotPredVsObs}()},
\code{\link{plotRatioVsCov}()},
\code{\link{plotResVsCov}()},
\code{\link{plotTimeProfile}()},
\code{\link{plotYVsX}()}
}
\concept{plot functions}
