% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRangeDistribution.R
\name{plotRangeDistribution}
\alias{plotRangeDistribution}
\title{Plot Range Plot}
\usage{
plotRangeDistribution(
  data,
  mapping,
  metaData = NULL,
  modeOfBinning = BINNINGMODE$number,
  numberOfBins = 20,
  breaks = NA,
  asStepPlot = FALSE,
  statFun = NULL,
  percentiles = getOspsuite.plots.option(optionKey = OptionKeys$Percentiles)[c(1, 3, 5)],
  yscale = "linear",
  yscale.args = list(),
  xscale = "linear",
  xscale.args = list(),
  geomRibbonAttributes = getDefaultGeomAttributes("Ribbon"),
  geomLineAttributes = getDefaultGeomAttributes("Line"),
  identifier = "IndividualId"
)
}
\arguments{
\item{data}{A data frame containing the data to be plotted. This data should include the variables specified in the \code{mapping} argument.}

\item{mapping}{A mapping object (created using \code{ggplot2::aes()}) that defines how variables in \code{data} are mapped to aesthetics such as x and y axes, color, fill, etc.}

\item{metaData}{Optional metadata to be added to the plot. This can include additional information relevant to the data being plotted.}

\item{modeOfBinning}{A character string specifying the mode of binning. It determines how the data will be divided into bins. Options include:
\itemize{
\item \verb{Equal Frequency Binning}
\item `Equal Width Binning
\item \verb{Custom Binning}
Default is \code{BINNINGMODE$number}.
}}

\item{numberOfBins}{An integer specifying the number of bins to use for equal frequency or width binning. Default is 20.}

\item{breaks}{Optional numeric vector specifying custom breaks for binning when \code{modeOfBinning} is set to \verb{Custom Binning}. This allows for precise control over how data is grouped.}

\item{asStepPlot}{A logical indicating whether to create a step plot. If TRUE, the plot will display steps between the data points rather than continuous lines. Default is FALSE.}

\item{statFun}{An optional function for statistical summary, which takes a vector of y-values and returns a summary (e.g., quantiles). If NULL, defaults to calculating quantiles based on the specified \code{percentiles}.}

\item{percentiles}{A numeric vector of percentiles to be used in the statistical summary, which defines the range of values to be displayed on the plot. Default is the 5th, 50th, and 95th percentiles.}

\item{yscale}{A character string specifying the y-axis scale. Options are "linear" or "log". This determines how the y values are displayed on the plot. Default is "linear".}

\item{yscale.args}{A list of additional arguments for the y-axis scale, which can be used to customize the appearance and behavior of the y-axis.}

\item{xscale}{A character string specifying the x-axis scale. Options are "linear" or "log". This determines how the x values are displayed on the plot. Default is "linear".}

\item{xscale.args}{A list of additional arguments for the x-axis scale, which can be used to customize the appearance and behavior of the x-axis.}

\item{geomRibbonAttributes}{A list of attributes for the ribbon geometry in the plot, allowing customization of the visual appearance, such as colors and transparency.}

\item{geomLineAttributes}{A list of attributes for the line geometry in the plot, allowing customization of line characteristics such as color, size, and type.}

\item{identifier}{columnName of individual identifiers, default "IndividualId"}
}
\value{
A ggplot object representing the range plot. The returned object can be further customized or rendered using \code{print()} or similar functions.
}
\description{
Creates a range plot using specified data and mapping, allowing for different binning strategies and scaling options. This function provides a flexible way to visualize data distributions over specified ranges with optional statistical summaries.
}
\seealso{
Other plot functions: 
\code{\link{plotBoxWhisker}()},
\code{\link{plotForest}()},
\code{\link{plotHistogram}()},
\code{\link{plotPredVsObs}()},
\code{\link{plotQQ}()},
\code{\link{plotRatioVsCov}()},
\code{\link{plotResVsCov}()},
\code{\link{plotTimeProfile}()},
\code{\link{plotYVsX}()}
}
\concept{plot functions}
