% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTimeProfile.R
\name{plotTimeProfile}
\alias{plotTimeProfile}
\title{generate time profile plots}
\usage{
plotTimeProfile(
  data = NULL,
  mapping = NULL,
  observedData = NULL,
  observedMapping = mapping,
  metaData = NULL,
  mapSimulatedAndObserved = NULL,
  xscale = AxisScales$linear,
  xscale.args = list(limits = c(0, NA)),
  yscale = AxisScales$linear,
  yscale.args = list(),
  y2scale = AxisScales$linear,
  y2scale.args = list(),
  plotObject = NULL,
  geomLineAttributes = getDefaultGeomAttributes("Line"),
  geomRibbonAttributes = getDefaultGeomAttributes("Ribbon"),
  geomPointAttributes = getDefaultGeomAttributes("Point"),
  geomErrorbarAttributes = getDefaultGeomAttributes("Errorbar"),
  geomLLOQAttributes = getDefaultGeomAttributes("LLOQ"),
  groupAesthetics = c("colour", "fill", "shape")
)
}
\arguments{
\item{data}{data.frame with simulated data will be displayed as lines with ribbons}

\item{mapping}{a list of aesthetic mappings to use for plot,
additional to \code{{ggplot2}} aesthetics, the aesthetics \code{groupby},\code{error},\code{error_relative},\code{lloq}, \code{mdv}, \code{y2axis} are available,
see vignettes for more details and examples}

\item{observedData}{data.frame with observed data  will be displayed as points with error-bars}

\item{observedMapping}{a list of aesthetic mappings to use for observed data,
per default is is set to mapping. So if both data sets have the same mapping, use only \code{mapping},
if a different mapping is necessary use \code{mapping} and \code{observedMapping}}

\item{metaData}{A named list of information about \code{data} such as the \code{dimension} and \code{unit} of its variables.}

\item{mapSimulatedAndObserved}{table with columns observed and simulated which maps simulated and observed data
use of \code{mapSimulatedAndObserved} triggers reset of aesthetic scales after simulation layers}

\item{xscale}{either 'linear' then \code{ggplot2::scale_x_continuous()} or 'log' then \code{ggplot2::scale_x_log10()} is used}

\item{xscale.args}{list of arguments passed to \code{ggplot2::scale_x_continuous()} or \code{ggplot2::scale_x_log10()}}

\item{yscale}{either 'linear' then \code{ggplot2::scale_y_continuous()} or 'log' then \code{ggplot2::scale_y_log10()} is used}

\item{yscale.args}{list of arguments passed to \code{ggplot2::scale_y_continuous()} or \code{ggplot2::scale_y_log10()}}

\item{y2scale}{either 'linear' the secondary axis is displayed linear, or 'log' secondary axis is displayed with log scale}

\item{y2scale.args}{list of arguments passed to \code{ggplot2::sec_axis()}, trans, break are set by code}

\item{plotObject}{An optional \code{ggplot} object on which to add the plot layers}

\item{geomLineAttributes}{A \code{list} with arguments which are passed on to the call \code{ggplot2::geom_line}}

\item{geomRibbonAttributes}{A \code{list} with arguments which are passed on to the call \code{ggplot2::geom_ribbon}}

\item{geomPointAttributes}{A \code{list} with arguments which are passed on to the call \code{ggplot2::geom_point}}

\item{geomErrorbarAttributes}{A \code{list} with arguments which are passed on to the call \code{ggplot2::geom_errorbar}}

\item{geomLLOQAttributes}{A \code{list} with arguments which are passed on to the call \code{ggplot2::geom_hline}}

\item{groupAesthetics}{vector of aesthetics, which are used for columns mapped with \code{groupby},}
}
\value{
A \code{ggplot} object
}
\description{
Produces time profiles for simulated and observed data.

For the simulated data a \code{geom_line} and a \code{geom_ribbon} layer are added
For the observed data a \code{geom_point} and a \code{geom_errorbar} layer are added

For more details and examples see the vignettes:
\itemize{
\item \code{vignette("Time Profile Plots", package = "ospsuite.plots")}
\item \code{vignette("ospsuite.plots", package = "ospsuite.plots")}
}
}
\examples{
\dontrun{
# Basic time profile plot with simulated data
plotTimeProfile(
  data = simulationData,
  mapping = aes(x = time, y = concentration, color = compound)
)

# Time profile with both simulated and observed data
plotTimeProfile(
  data = simulationData,
  observedData = observedData,
  mapping = aes(x = time, y = concentration, color = treatment),
  observedMapping = aes(x = time, y = concentration, color = treatment)
)

# Time profile with secondary y-axis
plotTimeProfile(
  data = myData,
  mapping = aes(x = time, y = concentration, y2axis = fraction_unbound)
)
}
}
\seealso{
Other plot functions: 
\code{\link{plotBoxWhisker}()},
\code{\link{plotForest}()},
\code{\link{plotHistogram}()},
\code{\link{plotPredVsObs}()},
\code{\link{plotQQ}()},
\code{\link{plotRatioVsCov}()},
\code{\link{plotResVsCov}()},
\code{\link{plotYVsX}()}
}
\concept{plot functions}
