% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotYVsX.R
\name{plotYVsX}
\alias{plotYVsX}
\title{Base Plot for Residuals and Predictions vs Covariates}
\usage{
plotYVsX(
  data,
  mapping,
  metaData = NULL,
  geomPointAttributes = getDefaultGeomAttributes("Point"),
  geomErrorbarAttributes = getDefaultGeomAttributes("Errorbar"),
  geomGuestLineAttributes = getDefaultGeomAttributes("GuestLine"),
  geomComparisonLineAttributes = getDefaultGeomAttributes("ComparisonLine"),
  geomLLOQAttributes = getDefaultGeomAttributes("LLOQ"),
  groupAesthetics = c("colour", "fill", "shape"),
  comparisonLineVector = NULL,
  addRegression = FALSE,
  addGuestLimits = FALSE,
  deltaGuest = 1,
  labelGuestCriteria = "guest criteria",
  residualScale = NULL,
  asSquarePlot = FALSE,
  xScale = AxisScales$linear,
  xScaleArgs = list(),
  yScale = AxisScales$log,
  yScaleArgs = list(),
  observedDataDirection = "y",
  yDisplayAsAbsolute = TRUE
)
}
\arguments{
\item{data}{A \code{data.frame} containing the data to plot.}

\item{mapping}{A list of aesthetic mappings to use for the plot.}

\item{metaData}{A named list of information about \code{data} such as the \code{dimension} and \code{unit} of its variables.}

\item{geomPointAttributes}{A \code{list} with arguments which are passed on to the call \code{ggplot2::geom_point}}

\item{geomErrorbarAttributes}{A \code{list} with arguments which are passed on to the call \code{ggplot2::geom_errorbar}}

\item{geomGuestLineAttributes}{A \code{list} of arguments passed to \code{ggplot2::geom_function} to display guest criteria.}

\item{geomComparisonLineAttributes}{A \code{list} of arguments passed to \code{ggplot2::hline} or \code{ggplot2::abline} to display comparison lines.}

\item{geomLLOQAttributes}{A \code{list} with arguments which are passed on to the call \code{ggplot2::geom_hline}}

\item{groupAesthetics}{A character vector of aesthetic names used for grouping data points when calculating
comparison statistics. Data will be grouped by combinations of these aesthetics before computing counts
and proportions within comparison lines. Common grouping aesthetics include \code{"colour"}, \code{"fill"}, \code{"shape"}.}

\item{comparisonLineVector}{A vector defining the comparison lines.}

\item{addRegression}{A boolean that activates the insertion of a regression line.}

\item{addGuestLimits}{A boolean that activates the insertion of guest limits.}

\item{deltaGuest}{Numeric value parameter for the Guest function.}

\item{labelGuestCriteria}{Label used in the legend for guest criteria (default: "guest criteria").}

\item{residualScale}{Either "linear", "log", or "ratio" scale for residuals.}

\item{asSquarePlot}{A boolean; if true, the plot is returned as a square plot with aspect ratio = 1 and fixed ratios.}

\item{xScale}{either 'linear' then \code{ggplot2::scale_x_continuous()} or 'log' then \code{ggplot2::scale_x_log10()} is used}

\item{xScaleArgs}{list of arguments passed to \code{ggplot2::scale_x_continuous()} or \code{ggplot2::scale_x_log10()}}

\item{yScale}{either 'linear' then \code{ggplot2::scale_y_continuous()} or 'log' then \code{ggplot2::scale_y_log10()} is used}

\item{yScaleArgs}{list of arguments passed to \code{ggplot2::scale_y_continuous()} or \code{ggplot2::scale_y_log10()}}

\item{observedDataDirection}{Either 'x' or 'y', defining the direction of observed data.}

\item{yDisplayAsAbsolute}{A boolean that defines the direction of comparison lines.}
}
\value{
A \code{ggplot} object representing the plotted data.
}
\description{
This function creates a base plot for \code{plotResVsCov()}, \code{plotRatioVsCov()}, and \code{plotPredVsObs()}.

For details and examples, see the vignettes:
\itemize{
\item \code{vignette("Goodness of fit", package = "ospsuite.plots")}
\item \code{vignette("ospsuite.plots", package = "ospsuite.plots")}
}
}
\seealso{
Other plot functions: 
\code{\link{plotBoxWhisker}()},
\code{\link{plotForest}()},
\code{\link{plotHistogram}()},
\code{\link{plotPredVsObs}()},
\code{\link{plotQQ}()},
\code{\link{plotRangeDistribution}()},
\code{\link{plotRatioVsCov}()},
\code{\link{plotResVsCov}()},
\code{\link{plotTimeProfile}()}
}
\concept{plot functions}
