% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotWithWatermark.R
\name{print.ggWatermark}
\alias{print.ggWatermark}
\title{Print method for ggWatermark objects}
\usage{
\method{print}{ggWatermark}(x, ...)
}
\arguments{
\item{x}{A ggWatermark object created by \code{ggplotWithWatermark()}.}

\item{...}{Additional arguments to be passed to the print method, allowing for further customization of the output.}
}
\value{
A ggplot object with a watermark drawn on it. The watermark is displayed according to the specified options.
}
\description{
This function customizes the printing of ggplot objects with the class "ggWatermark" by adding a watermark.
}
