% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-defaults.R
\name{resetDefaultColorMapDistinct}
\alias{resetDefaultColorMapDistinct}
\title{reset the default color map for discrete colors}
\usage{
resetDefaultColorMapDistinct(oldColorMaps)
}
\arguments{
\item{oldColorMaps}{list of color maps previously set}
}
\description{
reset the default color map for discrete colors
}
\seealso{
Other setDefault functions:
\code{\link{Shapes}},
\code{\link{colorMaps}},
\code{\link[=geom_point_osp]{geom_point_osp()}},
\code{\link[=getDefaultGeomAttributes]{getDefaultGeomAttributes()}},
\code{\link[=getDefaultOptions]{getDefaultOptions()}},
\code{\link[=getOspsuite.plots.option]{getOspsuite.plots.option()}},
\code{\link[=resetDefaultTheme]{resetDefaultTheme()}},
\code{\link[=resetDefaults]{resetDefaults()}},
\code{\link[=scale_shape_osp]{scale_shape_osp()}},
\code{\link[=scale_shape_osp_identity]{scale_shape_osp_identity()}},
\code{\link[=scale_shape_osp_manual]{scale_shape_osp_manual()}},
\code{\link[=setDefaultColorMapDistinct]{setDefaultColorMapDistinct()}},
\code{\link[=setDefaultTheme]{setDefaultTheme()}},
\code{\link[=setDefaults]{setDefaults()}},
\code{\link[=setOspsuite.plots.option]{setOspsuite.plots.option()}},
\code{\link[=stat_qq_osp]{stat_qq_osp()}}
}
\concept{setDefault functions}
