% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-defaults.R
\name{resetDefaults}
\alias{resetDefaults}
\title{restore to previously stored settings}
\usage{
resetDefaults(oldDefaults)
}
\arguments{
\item{oldDefaults}{\code{list}with previously stored settings}
}
\description{
restore to previously stored settings
}
\seealso{
Other setDefault functions:
\code{\link{colorMaps}},
\code{\link[=getDefaultGeomAttributes]{getDefaultGeomAttributes()}},
\code{\link[=getDefaultOptions]{getDefaultOptions()}},
\code{\link[=getOspsuite.plots.option]{getOspsuite.plots.option()}},
\code{\link[=resetDefaultColorMapDistinct]{resetDefaultColorMapDistinct()}},
\code{\link[=resetDefaultTheme]{resetDefaultTheme()}},
\code{\link[=setDefaultColorMapDistinct]{setDefaultColorMapDistinct()}},
\code{\link[=setDefaultTheme]{setDefaultTheme()}},
\code{\link[=setDefaults]{setDefaults()}},
\code{\link[=setOspsuite.plots.option]{setOspsuite.plots.option()}}
}
\concept{setDefault functions}
