% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_enum.R
\docType{data}
\name{ResidualScales}
\alias{ResidualScales}
\title{enumeration keys for OSPSuite.plots scaling options for residual calculations}
\format{
An object of class \code{list} of length 3.
}
\usage{
ResidualScales
}
\description{
enumeration keys for OSPSuite.plots scaling options for residual calculations
}
\keyword{datasets}
