% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_shapes.R
\name{scale_shape_osp}
\alias{scale_shape_osp}
\title{OSP Shape Scale}
\usage{
scale_shape_osp(...)
}
\arguments{
\item{...}{Passed to \code{ggplot2::discrete_scale}.}
}
\value{
A ggplot2 scale that can be added to a plot.
}
\description{
Discrete shape scale that automatically assigns shapes from
\code{ospShapeNames} in order based on the number of factor levels.
Equivalent to \code{ggplot2::scale_shape()}.

If there are more levels than available shapes, shapes are recycled
and a warning is issued.
}
\examples{
library(ggplot2)
df <- data.frame(x = 1:3, y = 1:3, group = c("A", "B", "C"))
ggplot(df, aes(x, y, shape = group)) +
  geom_point_osp(size = 4) +
  scale_shape_osp()
}
\seealso{
Other shapes:
\code{\link{Shapes}},
\code{\link{ospShapeNames}},
\code{\link[=scale_shape_osp_identity]{scale_shape_osp_identity()}},
\code{\link[=scale_shape_osp_manual]{scale_shape_osp_manual()}}
}
\concept{shapes}
