% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_shapes.R
\name{scale_shape_osp_manual}
\alias{scale_shape_osp_manual}
\title{OSP Shape Manual Scale}
\usage{
scale_shape_osp_manual(values, ...)
}
\arguments{
\item{values}{Named character vector. Names are factor levels;
values are entries from \code{ospShapeNames}.}

\item{...}{Passed to \code{ggplot2::scale_shape_manual}.}
}
\value{
A ggplot2 scale that can be added to a plot.
}
\description{
Manual shape scale for explicit mapping of factor levels to OSP shape names.
Equivalent to \code{ggplot2::scale_shape_manual()}.
}
\examples{
library(ggplot2)
df <- data.frame(x = 1:3, y = 1:3, group = c("A", "B", "C"))
ggplot(df, aes(x, y, shape = group)) +
  geom_point_osp(size = 4) +
  scale_shape_osp_manual(values = c(A = "circle", B = "diamond", C = "star"))
}
\seealso{
Other setDefault functions: 
\code{\link{Shapes}},
\code{\link{colorMaps}},
\code{\link{geom_point_osp}()},
\code{\link{getDefaultGeomAttributes}()},
\code{\link{getDefaultOptions}()},
\code{\link{getOspsuite.plots.option}()},
\code{\link{resetDefaultColorMapDistinct}()},
\code{\link{resetDefaultShapeDiscrete}()},
\code{\link{resetDefaultTheme}()},
\code{\link{resetDefaults}()},
\code{\link{scale_shape_osp}()},
\code{\link{scale_shape_osp_identity}()},
\code{\link{setDefaultColorMapDistinct}()},
\code{\link{setDefaultShapeDiscrete}()},
\code{\link{setDefaultTheme}()},
\code{\link{setDefaults}()},
\code{\link{setOspsuite.plots.option}()},
\code{\link{stat_qq_osp}()}
}
\concept{setDefault functions}
