% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-defaults.R
\name{setDefaults}
\alias{setDefaults}
\title{sets the defaults for the OSPSuite.plots package}
\usage{
setDefaults(defaultOptions = list(), colorMapList = NULL)
}
\arguments{
\item{defaultOptions}{list of options}

\item{colorMapList}{list of color maps}
}
\value{
list of old settings which can be used to reset defaults with \code{resetDefaults()}
}
\description{
should be started at the beginning at each workflow
}
\details{
for detailed information see
\code{vignette("ospsuite.plots", package = "ospsuite.plots")}
}
\seealso{
Other setDefault functions:
\code{\link{colorMaps}},
\code{\link[=getDefaultGeomAttributes]{getDefaultGeomAttributes()}},
\code{\link[=getDefaultOptions]{getDefaultOptions()}},
\code{\link[=getOspsuite.plots.option]{getOspsuite.plots.option()}},
\code{\link[=resetDefaultColorMapDistinct]{resetDefaultColorMapDistinct()}},
\code{\link[=resetDefaultTheme]{resetDefaultTheme()}},
\code{\link[=resetDefaults]{resetDefaults()}},
\code{\link[=setDefaultColorMapDistinct]{setDefaultColorMapDistinct()}},
\code{\link[=setDefaultTheme]{setDefaultTheme()}},
\code{\link[=setOspsuite.plots.option]{setOspsuite.plots.option()}},
\code{\link[=themeOspsuite]{themeOspsuite()}}
}
\concept{setDefault functions}
