% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-defaults.R
\name{setOspsuite.plots.option}
\alias{setOspsuite.plots.option}
\title{Set OSPSuite plots option with a given key and value.}
\usage{
setOspsuite.plots.option(optionKey, value)
}
\arguments{
\item{optionKey}{The key for the option.}

\item{value}{The value for the option.}
}
\description{
Set OSPSuite plots option with a given key and value.
}
\examples{
\dontrun{
setOspsuite.plots.option(optionKey = OptionKeys$watermarkEnabled, value = TRUE)
}
}
\seealso{
Other setDefault functions:
\code{\link{colorMaps}},
\code{\link[=getDefaultGeomAttributes]{getDefaultGeomAttributes()}},
\code{\link[=getDefaultOptions]{getDefaultOptions()}},
\code{\link[=getOspsuite.plots.option]{getOspsuite.plots.option()}},
\code{\link[=resetDefaultColorMapDistinct]{resetDefaultColorMapDistinct()}},
\code{\link[=resetDefaultTheme]{resetDefaultTheme()}},
\code{\link[=resetDefaults]{resetDefaults()}},
\code{\link[=setDefaultColorMapDistinct]{setDefaultColorMapDistinct()}},
\code{\link[=setDefaultTheme]{setDefaultTheme()}},
\code{\link[=setDefaults]{setDefaults()}}
}
\concept{setDefault functions}
