% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_shapes.R
\name{Shapes}
\alias{Shapes}
\title{Shapes}
\usage{
Shapes
}
\description{
Named list of OSP shape names for backward compatibility.
Use \code{Shapes$circle} to get the shape name "circle".
}
\seealso{
Other shapes:
\code{\link{ospShapeNames}},
\code{\link[=scale_shape_osp]{scale_shape_osp()}},
\code{\link[=scale_shape_osp_identity]{scale_shape_osp_identity()}},
\code{\link[=scale_shape_osp_manual]{scale_shape_osp_manual()}}
}
\concept{shapes}
