% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-defaults.R
\name{themeOspsuite}
\alias{themeOspsuite}
\title{OSPSuite plot theme}
\usage{
themeOspsuite(base_size = 11, ...)
}
\arguments{
\item{base_size}{base font size, given in pts (passed on to \code{\link[ggplot2:theme_bw]{ggplot2::theme_bw()}}).}

\item{...}{further arguments passed on to \code{\link[ggplot2:theme_bw]{ggplot2::theme_bw()}}
(for example \code{base_family}).}
}
\value{
a complete \code{ggplot2} theme object.
}
\description{
A \code{ggplot2} theme with OSPSuite-specific styling, based on
\code{\link[ggplot2:theme_bw]{ggplot2::theme_bw()}}. Apply it to a single plot with \code{plot + themeOspsuite()}
without altering the global \code{ggplot2} state.
}
\examples{
library(ggplot2)
ggplot(mtcars, aes(x = wt, y = mpg)) +
  geom_point() +
  themeOspsuite()

}
\seealso{
Other setDefault functions:
\code{\link{colorMaps}},
\code{\link[=getDefaultGeomAttributes]{getDefaultGeomAttributes()}},
\code{\link[=getDefaultOptions]{getDefaultOptions()}},
\code{\link[=getOspsuite.plots.option]{getOspsuite.plots.option()}},
\code{\link[=resetDefaultColorMapDistinct]{resetDefaultColorMapDistinct()}},
\code{\link[=resetDefaultTheme]{resetDefaultTheme()}},
\code{\link[=resetDefaults]{resetDefaults()}},
\code{\link[=setDefaultColorMapDistinct]{setDefaultColorMapDistinct()}},
\code{\link[=setDefaultTheme]{setDefaultTheme()}},
\code{\link[=setDefaults]{setDefaults()}},
\code{\link[=setOspsuite.plots.option]{setOspsuite.plots.option()}}
}
\concept{setDefault functions}
