% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_export.R
\name{validateFilename}
\alias{validateFilename}
\title{replace of special letters}
\usage{
validateFilename(filename)
}
\arguments{
\item{filename}{name of file to validate}
}
\value{
filename without special letters
}
\description{
replace of special letters
}
