% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_export.R
\name{validateInputsExportPlot}
\alias{validateInputsExportPlot}
\title{Validate inputs for exporting a plot}
\usage{
validateInputsExportPlot(plotObject, filepath, filename, width, height, device)
}
\arguments{
\item{plotObject}{A ggplot object to be exported.}

\item{filepath}{A character string specifying the directory to save the plot.}

\item{filename}{A character string specifying the name of the file (without path).}

\item{width}{A numeric value specifying the width of the plot.}

\item{height}{A numeric value specifying the height of the plot.}

\item{device}{A character with the device to use}
}
\description{
Validate inputs for exporting a plot
}
\keyword{internal}
