% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotYVsX.R
\name{addComparisonLines}
\alias{addComparisonLines}
\title{Add Comparison Lines to Plot}
\usage{
addComparisonLines(
  plotObject,
  comparisonLineVector,
  yDisplayAsAbsolute,
  geomLineAttributes,
  xyScale
)
}
\arguments{
\item{comparisonLineVector}{A vector defining the comparison lines.}

\item{yDisplayAsAbsolute}{A boolean that defines the direction of comparison lines.}

\item{geomLineAttributes}{Line attributes, e.g., \code{color}, \code{linetype}, passed to \code{ggplot2::geom_hline} or \code{ggplot2::geom_abline}.}

\item{xyScale}{Either "linear" or "log" scale for the X and Y axes.}
}
\value{
The updated \code{ggplot} object with comparison lines added.
}
\description{
This function adds horizontal or diagonal comparison lines to the given ggplot object.
}
\keyword{internal}
