% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_Layer.R
\name{addLayer}
\alias{addLayer}
\title{Add Layer}
\usage{
addLayer(mappedData, geomAttributes, geom, plotObject, layerToCall)
}
\arguments{
\item{mappedData}{A \code{MappedData} object.}

\item{geomAttributes}{Arguments passed on to the ggplot2 geom layer.}

\item{geom}{A character string used to select appropriate aesthetics.}

\item{plotObject}{A \code{ggplot} object on which to add the plot layer.}

\item{layerToCall}{A function representing the ggplot2 geom layer.}
}
\value{
The updated \code{ggplot} object.
}
\description{
Add a layer to a \code{ggplot} object.
}
\keyword{internal}
