% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_Layer.R
\name{addLLOQLayer}
\alias{addLLOQLayer}
\title{Add LLOQ Layer with LLOQ Lines}
\usage{
addLLOQLayer(
  plotObject,
  mappedData,
  layerToCall,
  useLinetypeAsAttribute,
  geomLLOQAttributes
)
}
\arguments{
\item{plotObject}{A \code{ggplot} object on which to add the plot layer.}

\item{mappedData}{A \code{MappedData} object with LLOQ data.}

\item{layerToCall}{A function representing the ggplot2 geom layer.}

\item{useLinetypeAsAttribute}{A boolean indicating whether to set the line type
as an attribute (TRUE) or not (FALSE); if TRUE, no legend is created.}

\item{geomLLOQAttributes}{Additional attributes for the LLOQ layer.}
}
\value{
The updated \code{ggplot} object.
}
\description{
Add a layer for LLOQ lines to a \code{ggplot} object.
}
