% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_Layer.R
\name{addXScale}
\alias{addXScale}
\title{add X-scale}
\usage{
addXScale(plotObject, xScale, xScaleArgs = list())
}
\arguments{
\item{plotObject}{A \code{ggplot} object on which to add the scale.}

\item{xScale}{The x-axis scale type. Available is 'linear', 'log', 'discrete'}

\item{xScaleArgs}{A list of arguments for the x-axis scale.}
}
\value{
The updated \code{ggplot} object
}
\description{
add X-scale
}
