% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_Layer.R
\name{addXscale}
\alias{addXscale}
\title{add X-scale}
\usage{
addXscale(plotObject, xscale, xscale.args = list())
}
\arguments{
\item{plotObject}{A \code{ggplot} object on which to add the scale.}

\item{xscale}{The x-axis scale type. Available is 'linear', 'log', 'discrete'}

\item{xscale.args}{A list of arguments for the x-axis scale.}
}
\value{
The updated \code{ggplot} object
}
\description{
add X-scale
}
