% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_Layer.R
\name{addXYScale}
\alias{addXYScale}
\title{Add X and Y Scale}
\usage{
addXYScale(
  plotObject,
  xscale = NULL,
  xscale.args = list(),
  yscale = NULL,
  yscale.args = list(),
  secAxis = waiver()
)
}
\arguments{
\item{plotObject}{A \code{ggplot} object on which to add the scale.}

\item{xscale}{The x-axis scale type. Available is 'linear', 'log', 'discrete'}

\item{xscale.args}{A list of arguments for the x-axis scale.}

\item{yscale}{The y-axis scale type. Available is 'linear', 'log'}

\item{yscale.args}{A list of arguments for the y-axis scale.}

\item{secAxis}{Secondary axis arguments for scale_y functions.}
}
\value{
The updated \code{ggplot} object.
}
\description{
Add X and Y scales to a \code{ggplot} object.
}
