% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_Layer.R
\name{addXYScale}
\alias{addXYScale}
\title{add X and Y-scale}
\usage{
addXYScale(
  plotObject,
  xscale = NULL,
  xscale.args = list(),
  yscale = NULL,
  yscale.args = list(),
  secAxis = waiver()
)
}
\arguments{
\item{plotObject}{A \code{ggplot} object on which to add the scale}

\item{xscale}{either 'linear' then \code{ggplot2::scale_x_continuous()} or 'log' then \code{ggplot2::scale_x_log10()} is used}

\item{xscale.args}{list of arguments passed to \code{ggplot2::scale_x_continuous()} or \code{ggplot2::scale_x_log10()}}

\item{yscale}{either 'linear' then \code{ggplot2::scale_y_continuous()} or 'log' then \code{ggplot2::scale_y_log10()} is used}

\item{yscale.args}{list of arguments passed to \code{ggplot2::scale_y_continuous()} or \code{ggplot2::scale_y_log10()}}

\item{secAxis}{secondary axis arguments for scale_y functions}
}
\value{
The updated \code{ggplot} object
}
\description{
add X and Y-scale
}
