% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_Layer.R
\name{addYscale}
\alias{addYscale}
\title{add y-scale}
\usage{
addYscale(plotObject, yscale, yscale.args = list(), secAxis = waiver())
}
\arguments{
\item{plotObject}{A \code{ggplot} object on which to add the scale.}

\item{yscale}{The y-axis scale type. Available is 'linear', 'log'}

\item{yscale.args}{A list of arguments for the y-axis scale.}

\item{secAxis}{Secondary axis arguments for scale_y functions.}
}
\value{
The updated \code{ggplot} object
}
\description{
add y-scale
}
