% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_Layer.R
\name{addYscale}
\alias{addYscale}
\title{add y-scale}
\usage{
addYscale(plotObject, yscale, yscale.args = list(), secAxis = waiver())
}
\arguments{
\item{plotObject}{A \code{ggplot} object on which to add the scale}

\item{yscale}{either 'linear' then ggplot2::scale_y_continuous() or 'log' then ggplot2::scale_y_log10() is used}

\item{yscale.args}{list of arguments passed to ggplot2::scale_y_continuous() or ggplot2::scale_y_log10()}

\item{secAxis}{description}
}
\value{
The updated \code{ggplot} object
}
\description{
add y-scale
}
