% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_Layer.R
\name{addYScale}
\alias{addYScale}
\title{add y-scale}
\usage{
addYScale(plotObject, yScale, yScaleArgs = list(), secAxis = waiver())
}
\arguments{
\item{plotObject}{A \code{ggplot} object on which to add the scale.}

\item{yScale}{The y-axis scale type. Available is 'linear', 'log'}

\item{yScaleArgs}{A list of arguments for the y-axis scale.}

\item{secAxis}{Secondary axis arguments for scale_y functions.}
}
\value{
The updated \code{ggplot} object
}
\description{
add y-scale
}
