% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_enum.R
\docType{data}
\name{AxisScales}
\alias{AxisScales}
\title{enumeration keys for OSPSuite.plots scaling options for axis scalings}
\format{
An object of class \code{list} of length 3.
}
\usage{
AxisScales
}
\description{
enumeration keys for OSPSuite.plots scaling options for axis scalings
}
\keyword{datasets}
