% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_export.R
\name{calculatePlotDimensions}
\alias{calculatePlotDimensions}
\title{Calculate plot dimensions based on the plot object and specified width}
\usage{
calculatePlotDimensions(plotObject, width)
}
\arguments{
\item{plotObject}{A ggplot object.}

\item{width}{A numeric value specifying the width of the plot.}
}
\value{
A list containing the calculated width, height, heightOffset, and widthOffset of the plot.
}
\description{
Calculate plot dimensions based on the plot object and specified width
}
\keyword{internal}
