% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-defaults.R
\docType{data}
\name{colorMaps}
\alias{colorMaps}
\title{Color maps}
\format{
An object of class \code{list} of length 6.
}
\usage{
colorMaps
}
\description{
List with some color maps for \code{Theme} object.

The \code{ospDefault} color map is based on colors in \code{default_igv} qualitative
color palette from \code{{ggsci}} package.
}
\seealso{
Other setDefault functions: 
\code{\link{Shapes}},
\code{\link{geom_point_osp}()},
\code{\link{getDefaultGeomAttributes}()},
\code{\link{getDefaultOptions}()},
\code{\link{getOspsuite.plots.option}()},
\code{\link{resetDefaultColorMapDistinct}()},
\code{\link{resetDefaultShapeDiscrete}()},
\code{\link{resetDefaultTheme}()},
\code{\link{resetDefaults}()},
\code{\link{scale_shape_osp}()},
\code{\link{scale_shape_osp_identity}()},
\code{\link{scale_shape_osp_manual}()},
\code{\link{setDefaultColorMapDistinct}()},
\code{\link{setDefaultShapeDiscrete}()},
\code{\link{setDefaultTheme}()},
\code{\link{setDefaults}()},
\code{\link{setOspsuite.plots.option}()},
\code{\link{stat_qq_osp}()}
}
\concept{setDefault functions}
\keyword{datasets}
