% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-defaults.R
\docType{data}
\name{ColorMaps}
\alias{ColorMaps}
\title{ColorMaps}
\format{
An object of class \code{list} of length 6.
}
\usage{
ColorMaps
}
\description{
List with some color maps for \code{Theme} object.

The \code{ospDefault} color map is based on colors in \code{default_igv} qualitative
color palette from \code{{ggsci}} package.
}
\seealso{
Other setDefaultsOspsuite.plots: 
\code{\link{getDefaultGeomAttributes}()},
\code{\link{getDefaultOptions}()},
\code{\link{resetDefaultColorMapDistinct}()},
\code{\link{resetDefaultShapeDiscrete}()},
\code{\link{resetDefaults}()},
\code{\link{setDefaultColorMapDistinct}()},
\code{\link{setDefaultShapeDiscrete}()},
\code{\link{setDefaultTheme}()},
\code{\link{setDefaults}()}
}
\concept{setDefaultsOspsuite.plots}
\keyword{datasets}
