% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{createDefaultPlotLabels}
\alias{createDefaultPlotLabels}
\title{create Default labels with unit for plot}
\usage{
createDefaultPlotLabels(metaData, mapping)
}
\arguments{
\item{metaData}{A named list of information about the \code{data} such as the \code{dimension} and \code{unit} of its variables.}

\item{mapping}{Default list of aesthetic mappings to use for plot}
}
\value{
list with plotLabels
}
\description{
create Default labels with unit for plot
}
\keyword{internal}
