% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{createDefaultPlotLabels}
\alias{createDefaultPlotLabels}
\title{create Default labels with unit for plot}
\usage{
createDefaultPlotLabels(mappedData)
}
\arguments{
\item{mappedData}{MappedData object with information of mapped dimensions and units}
}
\value{
list with plotLabels
}
\description{
create Default labels with unit for plot
}
\keyword{internal}
