% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotForest.R
\name{createPlotObject}
\alias{createPlotObject}
\title{Create the Plot Object}
\usage{
createPlotObject(
  plotData,
  mapping,
  xscale,
  xscale.args,
  xLabel,
  groupAesthetics,
  yFacetColumns,
  xFacetColumn,
  labelWrapWidth,
  geomPointAttributes,
  geomErrorbarAttributes,
  facetScales = "free_y"
)
}
\arguments{
\item{plotData}{A data.table containing the data to be plotted.}

\item{mapping}{A ggplot mapping object.}

\item{xscale}{A character string indicating the scale type for the x-axis.}

\item{xscale.args}{A list of additional arguments for customizing the x-axis scale.}

\item{xLabel}{A string representing the label for the x-axis.}

\item{groupAesthetics}{A character vector specifying aesthetics for grouping.}

\item{yFacetColumns}{A character vector of column names used for faceting on the y-axis.}

\item{xFacetColumn}{A character string specifying the column name for the x-axis facet.}

\item{labelWrapWidth}{A numeric value specifying the width for label wrapping in facets.}

\item{geomPointAttributes}{A list of attributes for the point geometry.}

\item{geomErrorbarAttributes}{A list of attributes for the error bar geometry.}

\item{facetScales}{A character string indicating the scales used for facets. Defaults to "free_y".}
}
\value{
A ggplot object representing the main plot.
}
\description{
This function generates the main plot object for the forest plot.
}
\keyword{internal}
