% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotForest.R
\name{createTableData}
\alias{createTableData}
\title{Create Table Data}
\usage{
createTableData(plotData, tableColumns, tableLabels)
}
\arguments{
\item{plotData}{A data.table containing the data to be plotted in the table.}

\item{tableColumns}{A character vector of column names to be included in the table.}

\item{tableLabels}{A character vector of labels corresponding to \code{tableColumns}.}
}
\value{
A data.table containing the prepared table data.
}
\description{
This function prepares the data for the table.
}
\keyword{internal}
