% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTimeProfile.R
\name{.addCommonLimitsAndYScaleArguments}
\alias{.addCommonLimitsAndYScaleArguments}
\title{set the common axis limits to \code{MappedData} object and update y2-scale arguments}
\usage{
.addCommonLimitsAndYScaleArguments(
  simMappedData,
  obsMappedData,
  commonLimits,
  requireDualAxis,
  yScaleArgs,
  y2ScaleArgs
)
}
\arguments{
\item{simMappedData}{object of class \code{MappedDataTimeprofile} for simulated data}

\item{obsMappedData}{object of class \code{MappedDataTimeprofile} for observed data}

\item{commonLimits}{common limits for simulated and observed data}

\item{requireDualAxis}{boolean if TRUE secondary axis is needed}

\item{yScaleArgs}{list with y2scale arguments}

\item{y2ScaleArgs}{list with arguments for secondary y-axis}
}
\value{
list with
\code{simMappedData} adjusted object of class \code{MappedDataTimeprofile} for simulated data
\code{obsMappedData} adjusted object of class \code{MappedDataTimeprofile} for observed data
\code{yScaleArgs} adjusted \code{yScaleArgs} with common limits for primary and secondary y-axis
\code{secAxis}  secondary axis object
}
\description{
set the common axis limits to \code{MappedData} object and update y2-scale arguments
}
\keyword{internal}
