% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTimeProfile.R
\name{.addCommonLimitsAndYscaleArguments}
\alias{.addCommonLimitsAndYscaleArguments}
\title{set the commonLimits to Mapped data and update y2ScaleARguments}
\usage{
.addCommonLimitsAndYscaleArguments(
  simMappedData,
  obsMappedData,
  commonLimits,
  requireDualAxis,
  yscale.args,
  y2scale.args
)
}
\arguments{
\item{simMappedData}{object of class \code{MappedDataTimeprofile} for simulated data}

\item{obsMappedData}{object of class \code{MappedDataTimeprofile} for observed data}

\item{commonLimits}{common limits for simulated and observed data}

\item{requireDualAxis}{boolena if TRUE secondary axis is needed}

\item{yscale.args}{list with y2scale arguments}

\item{y2scale.args}{list with arguments for secondary axisx}
}
\value{
list with
simMappedData adjusted object of class \code{MappedDataTimeprofile} for simulated data
obsMappedData adjusted object of class \code{MappedDataTimeprofile} for observed data
yscale.args adjusted yscale.args with common limits for primary and secondayr axisx
secAxis  secondary axis object
}
\description{
set the commonLimits to Mapped data and update y2ScaleARguments
}
\keyword{internal}
