% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTimeProfile.R
\name{.adjustLimits}
\alias{.adjustLimits}
\title{compares old and new limits}
\usage{
.adjustLimits(commonLimitsOld, commonLimitsNew)
}
\arguments{
\item{commonLimitsOld}{list of old limits}

\item{commonLimitsNew}{list of new limits}
}
\value{
updated limits
}
\description{
compares old and new limits
}
\keyword{internal}
