% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTimeProfile.R
\name{.initialplotObjectForTimeProfile}
\alias{.initialplotObjectForTimeProfile}
\title{initializes plot object and set scaling}
\usage{
.initialplotObjectForTimeProfile(
  simMappedData,
  obsMappedData,
  plotObject,
  xscale,
  xscale.args,
  yscale,
  yscale.args,
  y2scale.args,
  secAxis
)
}
\arguments{
\item{simMappedData}{object of class \code{MappedDataTimeprofile} for simulated data}

\item{obsMappedData}{object of class \code{MappedDataTimeprofile} for observed data}

\item{plotObject}{An optional \code{ggplot} object on which to add the plot layers}

\item{xscale}{either 'linear' then \code{ggplot2::scale_x_continuous()} or 'log' then \code{ggplot2::scale_x_log10()} is used}

\item{xscale.args}{list of arguments passed to \code{ggplot2::scale_x_continuous()} or \code{ggplot2::scale_x_log10()}}

\item{yscale}{either 'linear' then \code{ggplot2::scale_y_continuous()} or 'log' then \code{ggplot2::scale_y_log10()} is used}

\item{yscale.args}{list of arguments passed to \code{ggplot2::scale_y_continuous()} or \code{ggplot2::scale_y_log10()}}

\item{y2scale.args}{list of arguments passed to \code{ggplot2::sec_axis()}, trans, break are set by code}
}
\value{
plot object
}
\description{
initializes plot object and set scaling
}
\keyword{internal}
