% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exampleDataTimeProfile}
\alias{exampleDataTimeProfile}
\title{data frame which includes example data used in the vignette
\code{vignette("Time Profile Plots", package = "ospsuite.plots")}}
\format{
An object of class \code{data.frame} with 575 rows and 11 columns.
}
\usage{
exampleDataTimeProfile
}
\description{
data frame which includes example data used in the vignette
\code{vignette("Time Profile Plots", package = "ospsuite.plots")}
}
\keyword{datasets}
