% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotYVsX.R
\name{getCountsWithin}
\alias{getCountsWithin}
\title{Counts entries within specific limits}
\usage{
getCountsWithin(
  data,
  yColumn,
  xColumn = NULL,
  foldDistance = getFoldDistanceList(c(1.5, 2)),
  addGuestLimits = FALSE,
  deltaGuest = 1,
  groups = NULL
)
}
\arguments{
\item{data}{A \code{data.frame} with data to plot}

\item{yColumn}{column name for values to count}

\item{addGuestLimits}{A \code{Flag} which activates insertion of regression line}

\item{deltaGuest}{Numeric value parameter of Guest function}

\item{groups}{column names to group}
}
\value{
data table with summary
}
\description{
Counts entries within specific limits
}
