% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotYVsX.R
\name{getCountsWithin}
\alias{getCountsWithin}
\title{Count Entries Within Specific Limits}
\usage{
getCountsWithin(
  data,
  mapping,
  comparisonLineVector = getFoldDistanceList(c(1.5, 2)),
  addGuestLimits = FALSE,
  deltaGuest = 1,
  groups = NULL,
  yDisplayAsAbsolute
)
}
\arguments{
\item{data}{A \code{data.frame} containing the data to plot.}

\item{mapping}{A list of aesthetic mappings to use for the plot.}

\item{comparisonLineVector}{A vector defining the comparison lines.}

\item{addGuestLimits}{A boolean that activates the insertion of guest limits.}

\item{deltaGuest}{Numeric value parameter for the Guest function.}

\item{groups}{Column names for grouping the counts.}

\item{yDisplayAsAbsolute}{A boolean that defines the direction of comparison lines.}

\item{yColumn}{The name of the Y column for values to count.}

\item{xColumn}{The name of the X column for values to count.}
}
\value{
A data table summarizing the counts within the specified limits.
}
\description{
This function counts entries within specific limits defined by the comparison lines and guest limits.
}
