% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-defaults.R
\name{getDefaultGeomAttributes}
\alias{getDefaultGeomAttributes}
\title{get the defaults for the geom attributes used as defaults in plot functions
see \code{vignette("ospsuite.plots", package = "ospsuite.plots")} how to chnage defaults}
\usage{
getDefaultGeomAttributes(geom)
}
\arguments{
\item{geom}{type of geom to return attributes}
}
\value{
list with default attributes
}
\description{
get the defaults for the geom attributes used as defaults in plot functions
see \code{vignette("ospsuite.plots", package = "ospsuite.plots")} how to chnage defaults
}
\seealso{
Other setDefaultsOspsuite.plots: 
\code{\link{ColorMaps}},
\code{\link{getDefaultOptions}()},
\code{\link{resetDefaultColorMapDistinct}()},
\code{\link{resetDefaultshapeDiscrete}()},
\code{\link{resetDefaults}()},
\code{\link{setDefaultColorMapDistinct}()},
\code{\link{setDefaultTheme}()},
\code{\link{setDefaultshapeDiscrete}()},
\code{\link{setDefaults}()}
}
\concept{setDefaultsOspsuite.plots}
