% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-defaults.R
\name{getDefaultOptions}
\alias{getDefaultOptions}
\title{get list of default options}
\usage{
getDefaultOptions()
}
\value{
names list with default options
}
\description{
get list of default options
}
\seealso{
Other setDefault functions: 
\code{\link{Shapes}},
\code{\link{colorMaps}},
\code{\link{geom_point_osp}()},
\code{\link{getDefaultGeomAttributes}()},
\code{\link{getOspsuite.plots.option}()},
\code{\link{resetDefaultColorMapDistinct}()},
\code{\link{resetDefaultShapeDiscrete}()},
\code{\link{resetDefaultTheme}()},
\code{\link{resetDefaults}()},
\code{\link{scale_shape_osp}()},
\code{\link{scale_shape_osp_identity}()},
\code{\link{scale_shape_osp_manual}()},
\code{\link{setDefaultColorMapDistinct}()},
\code{\link{setDefaultShapeDiscrete}()},
\code{\link{setDefaultTheme}()},
\code{\link{setDefaults}()},
\code{\link{setOspsuite.plots.option}()},
\code{\link{stat_qq_osp}()}
}
\concept{setDefault functions}
