% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-defaults.R
\name{getDefaultOptions}
\alias{getDefaultOptions}
\title{get list of default options}
\usage{
getDefaultOptions()
}
\value{
names list with default options
}
\description{
get list of default options
}
\seealso{
Other setDefaultsOspsuite.plots: 
\code{\link{ColorMaps}},
\code{\link{getDefaultGeomAttributes}()},
\code{\link{resetDefaultColorMapDistinct}()},
\code{\link{resetDefaultshapeDiscrete}()},
\code{\link{resetDefaults}()},
\code{\link{setDefaultColorMapDistinct}()},
\code{\link{setDefaultTheme}()},
\code{\link{setDefaultshapeDiscrete}()},
\code{\link{setDefaults}()}
}
\concept{setDefaultsOspsuite.plots}
