% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotYVsX.R
\name{getGuestLimits}
\alias{getGuestLimits}
\title{calculates limits for DDI ratio}
\usage{
getGuestLimits(x, deltaGuest = 1, addLinesDiagnonal = FALSE, asLower = TRUE)
}
\arguments{
\item{x}{Numeric values input of Guest function}

\item{deltaGuest}{Numeric value parameter of Guest function}

\item{addLinesDiagnonal}{A \code{boolean}which defines direction of comparison lines}

\item{asLower}{function returns lower limit}
}
\value{
limit of guest function for x
}
\description{
calculates limits for DDI ratio
}
\references{
\url{https://dmd.aspetjournals.org/content/39/2/170}
}
\keyword{internal}
