% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-defaults.R
\name{getOspsuite.plots.option}
\alias{getOspsuite.plots.option}
\title{returns an option value for a option defined by the package OSPSuite.plots}
\usage{
getOspsuite.plots.option(optionKey)
}
\arguments{
\item{optionKey}{identifier of option}
}
\value{
option value
}
\description{
returns an option value for a option defined by the package OSPSuite.plots
}
\examples{
\dontrun{
# Set the option first before getting it
options(ospsuite.plots.watermark_enabled = TRUE)
getOspsuite.plots.option(optionKey = OptionKeys$watermark_enabled)
}
}
\seealso{
Other setDefault functions: 
\code{\link{Shapes}},
\code{\link{colorMaps}},
\code{\link{geomPointUnicode}()},
\code{\link{getDefaultGeomAttributes}()},
\code{\link{getDefaultOptions}()},
\code{\link{resetDefaultColorMapDistinct}()},
\code{\link{resetDefaultShapeDiscrete}()},
\code{\link{resetDefaultTheme}()},
\code{\link{resetDefaults}()},
\code{\link{setDefaultColorMapDistinct}()},
\code{\link{setDefaultShapeDiscrete}()},
\code{\link{setDefaultTheme}()},
\code{\link{setDefaults}()},
\code{\link{setOspsuite.plots.option}()}
}
\concept{setDefault functions}
