% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-defaults.R
\name{getOspsuite.plots.option}
\alias{getOspsuite.plots.option}
\title{returns an option value for a option defined by the package ospsuite.plots}
\usage{
getOspsuite.plots.option(optionKey)
}
\arguments{
\item{optionKey}{identifier of option}
}
\value{
option value
}
\description{
returns an option value for a option defined by the package ospsuite.plots
}
\examples{
getOspsuite.plots.option(optionKey = OptionKeys$watermark_enabled)
}
