% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_export.R
\name{getPlotDimensions}
\alias{getPlotDimensions}
\title{gets dimension of plots}
\usage{
getPlotDimensions(plotObject, exportunits, nCol, nRow, nPanel)
}
\arguments{
\item{plotObject}{ggplot object}

\item{exportunits}{units of exported figure}

\item{nCol}{for panel plots number of cols}

\item{nRow}{for panel plots number of cols}

\item{nPanel}{for panel plots number of panels}
}
\value{
list with dimension o plot components
}
\description{
gets dimension of plots
}
