% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_Layer.R
\name{initializePlot}
\alias{initializePlot}
\title{initialize Plot}
\usage{
initializePlot(mappedData = NULL, setMapping = TRUE)
}
\arguments{
\item{mappedData}{MappedData object}

\item{setMapping}{if TRUE (default) mapping is passed to ggplot, otherwise mapping will be used only to create labels}
}
\value{
A \code{ggplot} object
}
\description{
Initialize a \code{ggplot} object
with watermark
and set its labels by metaData
}
