% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_Layer.R
\name{initializePlot}
\alias{initializePlot}
\title{Initialize Plot}
\usage{
initializePlot(mappedData = NULL, setMapping = TRUE)
}
\arguments{
\item{mappedData}{A \code{MappedData} object.}

\item{setMapping}{A boolean indicating if TRUE (default) mapping is passed to ggplot; otherwise, mapping will be used only to create labels.}
}
\value{
A \code{ggplot} object.
}
\description{
Initialize a \code{ggplot} object with a watermark and set its labels by metaData.
}
