% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MappedDataBoxplot.R
\name{MappedDataBoxplot}
\alias{MappedDataBoxplot}
\title{MappedDataBoxplot}
\description{
R6 class for mapping variable to data for boxplot visualizations.
This class extends MappedData to provide specialized mapping functionality for box-and-whisker plots,
including handling of discrete and continuous x-axis scales and automatic grouping logic.
}
\examples{
\dontrun{
# Create boxplot mapping with continuous x variable
boxplotData <- MappedDataBoxplot$new(
  data = myDataFrame,
  mapping = aes(x = dose, y = concentration),
  xscale = "linear"
)

# Create boxplot mapping with categorical x variable
boxplotData <- MappedDataBoxplot$new(
  data = myDataFrame,
  mapping = aes(x = treatment_group, y = response),
  xscale = "discrete"
)
}
}
\seealso{
Other MappedData classes: 
\code{\link{MappedData}},
\code{\link{MappedDataTimeProfile}}
}
\concept{MappedData classes}
\section{Super class}{
\code{\link[ospsuite.plots:MappedData]{ospsuite.plots::MappedData}} -> \code{MappedDataBoxplot}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{xscale}}{scale of x axis}

\item{\code{xscale.args}}{arguments for scale of x axis}

\item{\code{hasXmapping}}{boolean, if TRUE x is mapped}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{boxwhiskerMapping}}{mapping for box whisker plot}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MappedDataBoxplot-new}{\code{MappedDataBoxplot$new()}}
\item \href{#method-MappedDataBoxplot-doAdjustmentsWithMetaData}{\code{MappedDataBoxplot$doAdjustmentsWithMetaData()}}
\item \href{#method-MappedDataBoxplot-clone}{\code{MappedDataBoxplot$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="ospsuite.plots" data-topic="MappedData" data-id="addMetaData"><a href='../../ospsuite.plots/html/MappedData.html#method-MappedData-addMetaData'><code>ospsuite.plots::MappedData$addMetaData()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ospsuite.plots" data-topic="MappedData" data-id="getAestheticsForGeom"><a href='../../ospsuite.plots/html/MappedData.html#method-MappedData-getAestheticsForGeom'><code>ospsuite.plots::MappedData$getAestheticsForGeom()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ospsuite.plots" data-topic="MappedData" data-id="updateScaleArgumentsForTimeUnit"><a href='../../ospsuite.plots/html/MappedData.html#method-MappedData-updateScaleArgumentsForTimeUnit'><code>ospsuite.plots::MappedData$updateScaleArgumentsForTimeUnit()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MappedDataBoxplot-new"></a>}}
\if{latex}{\out{\hypertarget{method-MappedDataBoxplot-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{MappedDataBoxplot} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MappedDataBoxplot$new(
  data,
  mapping,
  groupAesthetics = NULL,
  direction = "y",
  isObserved = TRUE,
  xlimits = NULL,
  ylimits = NULL,
  xscale = AxisScales$linear,
  yscale = AxisScales$linear,
  residualScale = NULL,
  residualAesthetic = "y"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{data.frame used for mapping}

\item{\code{mapping}}{list of aesthetic mappings}

\item{\code{groupAesthetics}}{vector of aesthetics, which are used for columns mapped with aesthetic \code{groupby}}

\item{\code{direction}}{direction of plot either "x" or "y"}

\item{\code{isObserved}}{A \code{boolean} if TRUE mappings mdv, lloq, error and error_relative are evaluated}

\item{\code{xlimits}}{limits for x-axis (may be NULL)}

\item{\code{ylimits}}{limits for y-axis (may be NULL)}

\item{\code{xscale}}{scale of x-axis either 'linear' or 'log'}

\item{\code{yscale}}{scale of y-axis either 'linear' or 'log'}

\item{\code{residualScale}}{scale of x residuals}

\item{\code{residualAesthetic}}{aesthetic used for mapping residuals}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{MappedDataBoxplot} class object
use Metadata to adjust binning of x-axis, and group aesthetic
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MappedDataBoxplot-doAdjustmentsWithMetaData"></a>}}
\if{latex}{\out{\hypertarget{method-MappedDataBoxplot-doAdjustmentsWithMetaData}{}}}
\subsection{Method \code{doAdjustmentsWithMetaData()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MappedDataBoxplot$doAdjustmentsWithMetaData(
  originalmapping,
  xscale,
  xscale.args
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{originalmapping}}{mapping provided by user}

\item{\code{xscale}}{either 'linear','log', 'discrete' or 'auto' (default) auto select linear for continuous data and discrete for categorical data}

\item{\code{xscale.args}}{list of arguments passed to \code{ggplot2::scale_x_continuous()}, \code{ggplot2::scale_x_log10()} or
\code{ggplot2::scale_x_discrete()}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
adjusted \code{MappedDataBoxplot} class object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MappedDataBoxplot-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MappedDataBoxplot-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MappedDataBoxplot$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
