% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MappedDataTimeProfile.R
\name{MappedDataTimeProfile}
\alias{MappedDataTimeProfile}
\title{MappedDataTimeProfile}
\description{
R6 class for mapping variable to \code{data}
}
\seealso{
Other MappedData classes: 
\code{\link{MappedData}},
\code{\link{MappedDataBoxplot}}
}
\concept{MappedData classes}
\section{Super class}{
\code{\link[ospsuite.plots:MappedData]{ospsuite.plots::MappedData}} -> \code{MappedDataTimeProfile}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{y2limits}}{double vector limits of secondary y axis}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{requireDualAxis}}{boolean, If TRUE secondary axis is required}

\item{\code{secAxis}}{sec_axis() object}

\item{\code{dataForPlot}}{scaled data used for plotting
adjust limits}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MappedDataTimeProfile-new}{\code{MappedDataTimeProfile$new()}}
\item \href{#method-MappedDataTimeProfile-scaleDataForSecondaryAxis}{\code{MappedDataTimeProfile$scaleDataForSecondaryAxis()}}
\item \href{#method-MappedDataTimeProfile-clone}{\code{MappedDataTimeProfile$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="ospsuite.plots" data-topic="MappedData" data-id="addMetaData"><a href='../../ospsuite.plots/html/MappedData.html#method-MappedData-addMetaData'><code>ospsuite.plots::MappedData$addMetaData()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ospsuite.plots" data-topic="MappedData" data-id="getAestheticsForGeom"><a href='../../ospsuite.plots/html/MappedData.html#method-MappedData-getAestheticsForGeom'><code>ospsuite.plots::MappedData$getAestheticsForGeom()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ospsuite.plots" data-topic="MappedData" data-id="updateScaleArgumentsForTimeUnit"><a href='../../ospsuite.plots/html/MappedData.html#method-MappedData-updateScaleArgumentsForTimeUnit'><code>ospsuite.plots::MappedData$updateScaleArgumentsForTimeUnit()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MappedDataTimeProfile-new"></a>}}
\if{latex}{\out{\hypertarget{method-MappedDataTimeProfile-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{MappedDataTimeProfile} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MappedDataTimeProfile$new(
  data,
  mapping,
  groupAesthetics = NULL,
  groupOrder = NULL,
  direction = "y",
  isObserved = TRUE,
  xlimits = NULL,
  ylimits = NULL,
  xscale = AxisScales$linear,
  scaleOfPrimaryAxis = AxisScales$linear,
  scaleOfSecondaryAxis = AxisScales$linear,
  y2limits = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{data.frame used for mapping}

\item{\code{mapping}}{list of aesthetic mappings}

\item{\code{groupAesthetics}}{vector of aesthetics, which are used for columns mapped with aesthetic \code{groupby} ,
use of group aesthetics triggers second axis after simulation layers}

\item{\code{groupOrder}}{labels and order for group aesthetic}

\item{\code{direction}}{direction of plot either "x" or "y"}

\item{\code{isObserved}}{A \code{boolean} if TRUE mappings mdv, lloq are evaluated}

\item{\code{xlimits}}{limits for x-axis (may be NULL)}

\item{\code{ylimits}}{limits for primary axis (may be NULL)}

\item{\code{xscale}}{= scale of x-axis}

\item{\code{scaleOfPrimaryAxis}}{scale of direction, either "linear" or "log"}

\item{\code{scaleOfSecondaryAxis}}{either 'linear' or 'log'}

\item{\code{y2limits}}{limits for secondary axis (may be NULL)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{MappedDataTimeProfile} object
scales data for secondary axis and updates \code{secAxis}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MappedDataTimeProfile-scaleDataForSecondaryAxis"></a>}}
\if{latex}{\out{\hypertarget{method-MappedDataTimeProfile-scaleDataForSecondaryAxis}{}}}
\subsection{Method \code{scaleDataForSecondaryAxis()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MappedDataTimeProfile$scaleDataForSecondaryAxis(
  ylimits = NULL,
  y2limits = NULL,
  y2scale.args = list()
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ylimits}}{limits for primary axis}

\item{\code{y2limits}}{limits for secondary axis}

\item{\code{y2scale.args}}{arguments for secondary axis}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
updated MappedDataTimeProfile
boolean for secondary axis
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MappedDataTimeProfile-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MappedDataTimeProfile-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MappedDataTimeProfile$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
