% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_enum.R
\docType{data}
\name{OptionKeys}
\alias{OptionKeys}
\title{enumeration keys for ospsuite.plots options}
\format{
An object of class \code{list} of length 22.
}
\usage{
OptionKeys
}
\description{
enumeration keys for ospsuite.plots options
}
\keyword{datasets}
