% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_shapes.R
\docType{data}
\name{ospShapeNames}
\alias{ospShapeNames}
\title{OSP Shape Names}
\format{
An object of class \code{character} of length 22.
}
\usage{
ospShapeNames
}
\description{
Character vector of all available OSP shape names.
}
\keyword{datasets}
