% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBoxWhisker.R
\name{plotBoxWhisker}
\alias{plotBoxWhisker}
\title{Title}
\usage{
plotBoxWhisker(
  data,
  mapping = NULL,
  metaData = NULL,
  plotObject = NULL,
  percentiles = getOption(x = "ospsuite.plots.Percentiles", default =
    getDefaultggOSPOptions()[["ospsuite.plots.Percentiles"]]),
  yscale = "linear",
  yscale.args = list(),
  xscale = "auto",
  xscale.args = list(),
  statFun = NULL,
  outliers = FALSE,
  statFunOutlier = NULL,
  geomBoxplotAttributes = getDefaultGeomAttributes("Boxplot"),
  geomPointAttributes = getDefaultGeomAttributes("Boxplot")
)
}
\arguments{
\item{plotObject}{}
}
\description{
Title
}
