% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBoxWhisker.R
\name{plotBoxWhisker}
\alias{plotBoxWhisker}
\title{generate  box-whisker plots}
\usage{
plotBoxWhisker(
  data,
  mapping,
  metaData = NULL,
  plotObject = NULL,
  percentiles = getOspsuite.plots.option(optionKey = OptionKeys$Percentiles),
  yscale = AxisScales$linear,
  yscale.args = list(),
  xscale = "auto",
  xscale.args = list(),
  statFun = NULL,
  outliers = FALSE,
  statFunOutlier = NULL,
  geomBoxplotAttributes = getDefaultGeomAttributes("Boxplot"),
  geomPointAttributes = getDefaultGeomAttributes("Boxplot"),
  residualScale = ResidualScales$log
)
}
\arguments{
\item{data}{data.frame with data to aggregate}

\item{mapping}{a list of aesthetic mappings to use for plot}

\item{metaData}{A named list of information about \code{data} such as the \code{dimension} and \code{unit} of its variables.}

\item{plotObject}{An optional \code{ggplot} object on which to add the plot layers}

\item{percentiles}{vector with percentiles used for the box whiskers and boxes e.g. c(0.05,0.25,0.5,0.75,0.95)
default defined by ospsuite.plots option}

\item{yscale}{either 'linear' then \code{ggplot2::scale_y_continuous()} or 'log' then \code{ggplot2::scale_y_log10()} is used}

\item{yscale.args}{list of arguments passed to \code{ggplot2::scale_y_continuous()} or \code{ggplot2::scale_y_log10()}}

\item{xscale}{either 'linear','log', discrete or 'auto' (default) auto select linear for continuous data and discrete for categorical data}

\item{xscale.args}{list of arguments passed to \code{ggplot2::scale_x_continuous()}, \code{ggplot2::scale_x_log10()} or
\code{ggplot2::scale_x_discrete()}}

\item{statFun}{(default NULL) if not NULL function to calculate whiskers and box ranges, overwrites variable percentiles}

\item{outliers}{Logical defining if outliers should be included in boxplot.
outliers are flagged when outside the range from "25th" percentile - 1.5 x IQR to "75th"
percentiles + 1.5 x IQR, as suggested by McGill and}

\item{statFunOutlier}{(default NULL) if not NULL overwrites default calculation of outliers}

\item{geomBoxplotAttributes}{A \code{list} with arguments which are passed on to the geom boxplot call}

\item{geomPointAttributes}{A \code{list} with arguments which are passed on to the call \code{ggplot2::geom_point}}

\item{residualScale}{either "linear","log" or "ratio" scale residuals,
* linear:  residuals are calculated observed - predicted
* log: residuals are calculated log(observed) - log(predicted)
* ratio: residuals are calculated as observed/predicted}
}
\value{
A \code{ggplot} object
}
\description{
Producing box-and-whisker plots
For more details and examples see the vignettes:
\itemize{
\item \code{vignette("Box-Whisker Plots", package = "ospsuite.plots")}
\item \code{vignette("ospsuite.plots", package = "ospsuite.plots")}
}
}
\seealso{
Other plot functions: 
\code{\link{plotHistogram}()},
\code{\link{plotPredVsObs}()},
\code{\link{plotQQ}()},
\code{\link{plotRatioVsCov}()},
\code{\link{plotResVsCov}()},
\code{\link{plotTimeProfile}()},
\code{\link{plotYVsX}()}
}
\concept{plot functions}
