% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBoxWhisker.R
\name{plotBoxWhisker}
\alias{plotBoxWhisker}
\title{Generate Box-Whisker Plots}
\usage{
plotBoxWhisker(
  data,
  mapping,
  metaData = NULL,
  plotObject = NULL,
  percentiles = getOspsuite.plots.option(optionKey = OptionKeys$Percentiles),
  yscale = AxisScales$linear,
  yscale.args = list(),
  xscale = "auto",
  xscale.args = list(),
  statFun = NULL,
  outliers = FALSE,
  statFunOutlier = NULL,
  geomBoxplotAttributes = getDefaultGeomAttributes("Boxplot"),
  geomPointAttributes = getDefaultGeomAttributes("Boxplot"),
  residualScale = ResidualScales$log
)
}
\arguments{
\item{data}{A \code{data.frame} containing the data to aggregate.}

\item{mapping}{A list of aesthetic mappings to use for the plot.}

\item{metaData}{A named list of information about \code{data} such as the \code{dimension} and \code{unit} of its variables.}

\item{plotObject}{An optional \code{ggplot} object on which to add the plot layers}

\item{percentiles}{A numeric vector with percentiles used for the box whiskers and boxes,
e.g., c(0.05, 0.25, 0.5, 0.75, 0.95). Default defined by \code{ospsuite.plots} option.}

\item{yscale}{either 'linear' then \code{ggplot2::scale_y_continuous()} or 'log' then \code{ggplot2::scale_y_log10()} is used}

\item{yscale.args}{list of arguments passed to \code{ggplot2::scale_y_continuous()} or \code{ggplot2::scale_y_log10()}}

\item{xscale}{Either 'linear', 'log', 'discrete', or 'auto' (default).
Auto selects linear for continuous data and discrete for categorical data.}

\item{xscale.args}{A list of arguments passed to \code{ggplot2::scale_x_continuous()},
\code{ggplot2::scale_x_log10()}, or \code{ggplot2::scale_x_discrete()}.}

\item{statFun}{(default NULL) A function to calculate whiskers and box ranges,
which overwrites the \code{percentiles} variable if provided.}

\item{outliers}{Logical indicating whether outliers should be included in the boxplot.
Outliers are flagged when outside the range from the "25th" percentile - 1.5 x IQR to
the "75th" percentile + 1.5 x IQR, as suggested by McGill et al.}

\item{statFunOutlier}{(default NULL) A function to calculate outliers,
which overwrites the default calculation if provided.}

\item{geomBoxplotAttributes}{A \code{list} of arguments passed to the \code{geom_boxplot} call.}

\item{geomPointAttributes}{A \code{list} of arguments passed to the \code{ggplot2::geom_point} call.}

\item{residualScale}{Either "linear", "log", or "ratio" scale for residuals.}
}
\value{
A \code{ggplot} object representing the box-whisker plot.
}
\description{
Produces box-and-whisker plots for visualizing the distribution of data.
For more details and examples, see the vignettes:
\itemize{
\item \code{vignette("Box-Whisker Plots", package = "ospsuite.plots")}
\item \code{vignette("ospsuite.plots", package = "ospsuite.plots")}
}
}
\references{
McGill, R., Tukey, J. W., & Larsen, W. A. (1978). Variations of box plots.
The American Statistician, 32(1), 12-16.
}
\seealso{
Other plot functions: 
\code{\link{plotHistogram}()},
\code{\link{plotPredVsObs}()},
\code{\link{plotQQ}()},
\code{\link{plotRatioVsCov}()},
\code{\link{plotResVsCov}()},
\code{\link{plotTimeProfile}()},
\code{\link{plotYVsX}()}
}
\concept{plot functions}
