% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHistogram.R
\name{plotHistogram}
\alias{plotHistogram}
\title{generates histograms}
\usage{
plotHistogram(
  data,
  mapping,
  metaData = NULL,
  asBarPlot = NULL,
  geomHistAttributes = getDefaultGeomAttributes("Hist"),
  plotAsFrequency = FALSE,
  xscale = "linear",
  xscale.args = list(),
  yscale = "linear",
  yscale.args = list(),
  distribution = "none",
  meanFunction = "auto",
  residualScale = "log"
)
}
\arguments{
\item{data}{data.frame with simulated data will be displayed as lines with ribbons}

\item{mapping}{a list of aesthetic mappings to use for plot,
additional to \code{{ggplot2}} aesthetics, the aesthetics \code{groupby},\code{error},\code{error_relative},\code{lloq}, \code{mdv}, \code{y2axis} are available,
see vignettes for more details and examples}

\item{metaData}{A named list of information about \code{data} such as the \code{dimension} and \code{unit} of its variables.}

\item{asBarPlot}{A \code{boolean} defining if geom_histogram should be used (continuous data) or geom_bar (categorical)
if TRUE variables \code{distribution}, \code{meanFunction}, \code{xacale} and \code{xscale.args} are ignored}

\item{geomHistAttributes}{A \code{list} with arguments which are passed on to the ggplot2::geom_histogram (geom_bar if \code{asBarPlot} = TRUE)}

\item{plotAsFrequency}{A \code{boolean} defining if histogram displays a frequency in y axis}

\item{xscale}{either 'linear' then \code{ggplot2::scale_x_continuous()} or 'log' then \code{ggplot2::scale_x_log10()} is used}

\item{xscale.args}{list of arguments passed to \code{ggplot2::scale_x_continuous()} or \code{ggplot2::scale_x_log10()}}

\item{yscale}{either 'linear' then \code{ggplot2::scale_y_continuous()} or 'log' then \code{ggplot2::scale_y_log10()} is used}

\item{yscale.args}{list of arguments passed to \code{ggplot2::scale_y_continuous()} or \code{ggplot2::scale_y_log10()}}

\item{distribution}{name of distribution  available are all distributions available in package stats: (see ? stats::distributions) \code{norm},\code{lnorm}, ...
accepted are also normal (instead of \code{norm}) and lognormal (instead of \code{lnorm})}

\item{meanFunction}{function selection for display of vertical line. Available are 'none','mean','geomean','median','auto' (default)
'auto' selects 'mean' for \code{normal} distribution, 'geomean' for \code{lognormal} distribution, 'median' for all other distributions, and 'none' for 'none' distribution}

\item{residualScale}{either "linear","log" or "ratio" scale residuals,
* linear:  residuals are calculated observed - predicted
* log: residuals are calculated log(observed) - log(predicted)
* ratio: residuals are calculated as observed/predicted}
}
\value{
A \code{ggplot} object
}
\description{
Produces histograms with optional distribution fit

For more details and examples see the vignettes:
\itemize{
\item \code{vignette("Histogram Plots", package = "ospsuite.plots")}
\item \code{vignette("ospsuite.plots", package = "ospsuite.plots")}
\item \code{vignette("Goodness of fit", package = "ospsuite.plots")}
}
}
\seealso{
Other plot functions: 
\code{\link{plotBoxWhisker}()},
\code{\link{plotPredVsObs}()},
\code{\link{plotQQ}()},
\code{\link{plotRatioVsCov}()},
\code{\link{plotResVsCov}()},
\code{\link{plotTimeProfile}()},
\code{\link{plotYVsX}()}
}
\concept{plot functions}
