% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotYVsX.R
\name{plotPredVsObs}
\alias{plotPredVsObs}
\title{generates predicted vs observed plots}
\usage{
plotPredVsObs(
  data = NULL,
  mapping = NULL,
  xyscale = "log",
  comparisonLineVector = getFoldDistanceList(c(1.5, 2)),
  asSquarePlot = TRUE,
  ...
)
}
\arguments{
\item{data}{A \code{data.frame} with data to plot}

\item{mapping}{a list of aesthetic mappings to use for plot}

\item{xyscale}{either "linear" or "log" scale of x and y axis}

\item{comparisonLineVector}{either a double vector or a list of double values
if add \code{addLinesDiagnonal = FALSE} lines will be added as horizontal lines with the
intercept at values of \code{comparisonLineVector}
If add \code{addLinesDiagnonal = TRUE} lines will be added as fold distance lines to the identity.}

\item{asSquarePlot}{A \code{boolean} if true plot is returned as square plot with aspect.ratio = 1 and fixed ratios}

\item{...}{
  Arguments passed on to \code{\link[=plotYVsX]{plotYVsX}}
  \describe{
    \item{\code{geomComparisonLineAttributes}}{A \code{list} with arguments which are passed on to the
call \code{ggplot2::hline}  or \code{ggplot2::abline} to display comparison lines}
    \item{\code{geomGuestLineAttributes}}{A \code{list} with arguments which are passed on to the
call \code{ggplot2::geom_function} to display guest criteria}
    \item{\code{residualScale}}{either "linear","log" or "ratio" scale residuals,
* linear:  residuals are calculated observed - predicted
* log: residuals are calculated log(observed) - log(predicted)
* ratio: residuals are calculated as observed/predicted}
    \item{\code{addLinesDiagnonal}}{A \code{boolean}which defines direction of comparison lines}
    \item{\code{addRegression}}{A \code{boolean} which activates insertion of regression line}
    \item{\code{addGuestLimits}}{A \code{boolean} which activates insertion of regression line}
    \item{\code{deltaGuest}}{Numeric value parameter of Guest function}
    \item{\code{observedDataDirection}}{either 'x' or 'y', defines direction of observed data. relevant for
aesthetics \code{lloq}, \code{error}end \code{error_relative}}
    \item{\code{metaData}}{A named list of information about \code{data} such as the \code{dimension} and \code{unit} of its variables.}
    \item{\code{geomLineAttributes}}{A \code{list} with arguments which are passed on to the call \code{ggplot2::geom_line}}
    \item{\code{geomPointAttributes}}{A \code{list} with arguments which are passed on to the call \code{ggplot2::geom_point}}
    \item{\code{geomErrorbarAttributes}}{A \code{list} with arguments which are passed on to the call \code{ggplot2::geom_errorbar}}
    \item{\code{geomLLOQAttributes}}{A \code{list} with arguments which are passed on to the call \code{ggplot2::geom_hline}}
    \item{\code{groupAesthetics}}{vector of aesthetics, which are used for columns mapped with \code{groupby},}
    \item{\code{xscale}}{either 'linear' then \code{ggplot2::scale_x_continuous()} or 'log' then \code{ggplot2::scale_x_log10()} is used}
    \item{\code{xscale.args}}{list of arguments passed to \code{ggplot2::scale_x_continuous()} or \code{ggplot2::scale_x_log10()}}
    \item{\code{yscale}}{either 'linear' then \code{ggplot2::scale_y_continuous()} or 'log' then \code{ggplot2::scale_y_log10()} is used}
    \item{\code{yscale.args}}{list of arguments passed to \code{ggplot2::scale_y_continuous()} or \code{ggplot2::scale_y_log10()}}
  }}
}
\value{
A \code{ggplot} object
}
\description{
This functions is a wrapper for function \code{plotYVsX} with adjusted input parameter.
\itemize{
\item \code{residualScale} is fixed to NULL,
\item \code{observedDataDirection} is fixed to 'x'
}

For details and examples see the vignettes:
\itemize{
\item \code{vignette("Goodness of fit", package = "ospsuite.plots")}
\item \code{vignette("ospsuite.plots", package = "ospsuite.plots")}
}
}
\seealso{
Other plot functions: 
\code{\link{plotBoxWhisker}()},
\code{\link{plotHistogram}()},
\code{\link{plotQQ}()},
\code{\link{plotRatioVsCov}()},
\code{\link{plotResVsCov}()},
\code{\link{plotTimeProfile}()},
\code{\link{plotYVsX}()}
}
\concept{plot functions}
