% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotQQ.R
\name{plotQQ}
\alias{plotQQ}
\title{generates residual quantile quantile plot}
\usage{
plotQQ(
  data = NULL,
  mapping = NULL,
  xscale.args = list(),
  residualScale = "log",
  yscale.args = list(),
  geomQQAttributes = list(),
  geomQQLineAttributes = geomQQAttributes,
  groupAesthetics = c("colour", "fill", "shape"),
  addRegression = FALSE
)
}
\arguments{
\item{data}{´data.frame` with data to plot}

\item{mapping}{a list of aesthetic mappings to use for plot,
additional to {ggplot2} aesthetics, the aesthetics \code{groupby},\code{error},\code{error_relative},\code{lloq}, \code{mdv}, \code{y2axis} are available,
see vignettes for more details and examples}

\item{xscale.args}{list of arguments passed to \code{ggplot2::scale_x_continuous()} or \code{ggplot2::scale_x_log10()}}

\item{residualScale}{either "linear" or "log" scale residuals,
for linear residuals are calculated observed - predicted
for log residuals are calculated log(observed) - log(predicted)
y scale is always linear}

\item{yscale.args}{list of arguments passed to \code{ggplot2::scale_y_continuous()} or \code{ggplot2::scale_y_log10()}}

\item{geomQQAttributes}{arguments passed on to ggplot2::qq_stat}

\item{geomQQLineAttributes}{arguments passed on to ggplot2::qq_stat_line}

\item{groupAesthetics}{vector of aesthetics, which are used for columns mapped with groupby,}
}
\value{
A \code{ggplot} object
}
\description{
For details and examples see the vignettes:
\itemize{
\item \code{vignette("Goodness of fit", package = "ospsuite.plots")}
\item \code{vignette("ospsuite.plots", package = "ospsuite.plots")}
}
}
\seealso{
Other plot functions: 
\code{\link{plotBoxWhisker}()},
\code{\link{plotHistogram}()},
\code{\link{plotPredVsObs}()},
\code{\link{plotRatioVsCov}()},
\code{\link{plotResVsCov}()},
\code{\link{plotTimeProfile}()},
\code{\link{plotYVsX}()}
}
\concept{plot functions}
