% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTimeProfile.R
\name{plotTimeProfile}
\alias{plotTimeProfile}
\title{generate time profile plots}
\usage{
plotTimeProfile(
  data = NULL,
  mapping = NULL,
  observedData = NULL,
  observedMapping = mapping,
  metaData = NULL,
  xscale = "linear",
  xscale.args = list(limits = c(0, NA)),
  yscale = "linear",
  yscale.args = list(),
  y2scale = "linear",
  y2scale.args = list(),
  plotObject = NULL,
  geomLineAttributes = getDefaultGeomAttributes("Line"),
  geomRibbonAttributes = getDefaultGeomAttributes("Ribbon"),
  geomPointAttributes = getDefaultGeomAttributes("Point"),
  geomErrorbarAttributes = getDefaultGeomAttributes("Errorbar"),
  geomLLOQAttributes = getDefaultGeomAttributes("LLOQ"),
  groupAesthetics = c("colour", "fill", "linetype", "shape")
)
}
\arguments{
\item{data}{data.frame with simulated data}

\item{mapping}{Default list of aesthetic mappings to use for plot}

\item{observedData}{data.frame with observed data}

\item{observedMapping}{Default list of aesthetic mappings to use for observed data}

\item{metaData}{A named list of information about \code{data} such as the \code{dimension} and \code{unit} of its variables.}

\item{xscale}{either 'linear' then ggplot2::scale_x_continuous() or 'log' then ggplot2::scale_x_log10() is used}

\item{xscale.args}{list of arguments passed to ggplot2::scale_x_continuous() or ggplot2::scale_x_log10()}

\item{yscale}{either 'linear' then ggplot2::scale_y_continuous() or 'log' then ggplot2::scale_y_log10() is used}

\item{yscale.args}{list of arguments passed to ggplot2::scale_y_continuous() or ggplot2::scale_y_log10()}

\item{y2scale}{either 'linear' the secondary axis is displayed linear, or 'log' secondary axis is displayed with log scale}

\item{y2scale.args}{list of arguments passed to ggplot2::sec_axis(), trans, break are set by code}

\item{plotObject}{An optional \code{ggplot} object on which to add the plot layers}

\item{geomLineAttributes}{A \code{list} with arguments which are passed on to the ggplot2::geom_line}

\item{geomRibbonAttributes}{A \code{list} with arguments which are passed on to the ggplot2::geom_ribbon}

\item{geomPointAttributes}{A \code{list} with arguments which are passed on to the ggplot2::geom_point}

\item{geomErrorbarAttributes}{A \code{list} with arguments which are passed on to the ggplot2::geom_errorbar}

\item{geomLLOQAttributes}{A \code{list} with arguments which are passed on to the ggplot2::geom_hline}

\item{groupAesthetics}{vector of aesthetics, which are used for columns mapped with group,
use of group aesthetics triggers second axis after simulation layers}
}
\value{
A \code{ggplot} object
}
\description{
Produces timeprofiles for simulated and observed data.
}
\details{
For the simulated data a geom_line and a geom_ribbon layer are added
For the observed data a geom_point and a geom_errorbar layer are added

As one of the intention of this plotting routine is to display simulated
and observed data with same grouping aesthetics like color and line type, but with different legend,
between the layers for simulated data and layers for observered  data new color scales are introduced.

Non default mappings which can be used are:
\itemize{
\item lloq
\itemize{
\item only relevant for observed data
\item adds horizontal lines and displays measurement below LLOQ with a smaller alpha)
}
\item mdv
\itemize{
\item missing dependent Variable , should be 0 or 1
\item only relevant for observed data,
\item data with mdv 1 are not plotted
}
\item error, error_relative
\itemize{
\item calculates ymin and ymax for additive ore relative
}
\item y2axis
\itemize{
\item should be logical, defines if some of the outputs should be displayed wit a secondary axis
}
}

For more details and examples see the vignettes:
\code{vignette("Time Profile Plots", package = "ospsuite.plots")}
\code{vignette("ospsuite.plots", package = "ospsuite.plots")}
}
\seealso{
Other plot functions: 
\code{\link{plotHistogram}()}
}
\concept{plot functions}
