% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-defaults.R
\name{resetDefaultColorMapDistinct}
\alias{resetDefaultColorMapDistinct}
\title{reset the default colormap for discret colors}
\usage{
resetDefaultColorMapDistinct(oldColorMaps)
}
\arguments{
\item{oldColorMaps}{list of colormaps prevoisuly set}
}
\description{
reset the default colormap for discret colors
}
\seealso{
Other setDefaultsOspsuite.plots: 
\code{\link{ColorMaps}},
\code{\link{getDefaultGeomAttributes}()},
\code{\link{getDefaultOptions}()},
\code{\link{resetDefaultshapeDiscrete}()},
\code{\link{resetDefaults}()},
\code{\link{setDefaultColorMapDistinct}()},
\code{\link{setDefaultTheme}()},
\code{\link{setDefaultshapeDiscrete}()},
\code{\link{setDefaults}()}
}
\concept{setDefaultsOspsuite.plots}
