% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-defaults.R
\name{resetDefaultColorMapDistinct}
\alias{resetDefaultColorMapDistinct}
\title{reset the default color map for discrete colors}
\usage{
resetDefaultColorMapDistinct(oldColorMaps)
}
\arguments{
\item{oldColorMaps}{list of color maps previously set}
}
\description{
reset the default color map for discrete colors
}
\seealso{
Other setDefault functions: 
\code{\link{Shapes}},
\code{\link{colorMaps}},
\code{\link{geomPointUnicode}()},
\code{\link{getDefaultGeomAttributes}()},
\code{\link{getDefaultOptions}()},
\code{\link{getOspsuite.plots.option}()},
\code{\link{resetDefaultShapeDiscrete}()},
\code{\link{resetDefaultTheme}()},
\code{\link{resetDefaults}()},
\code{\link{setDefaultColorMapDistinct}()},
\code{\link{setDefaultShapeDiscrete}()},
\code{\link{setDefaultTheme}()},
\code{\link{setDefaults}()},
\code{\link{setOspsuite.plots.option}()}
}
\concept{setDefault functions}
