% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-defaults.R
\name{resetDefaultColorMapDistinct}
\alias{resetDefaultColorMapDistinct}
\title{reset the default colormap for discrete colors}
\usage{
resetDefaultColorMapDistinct(oldColorMaps)
}
\arguments{
\item{oldColorMaps}{list of colormaps previously set}
}
\description{
reset the default colormap for discrete colors
}
\seealso{
Other setDefault functions: 
\code{\link{ColorMaps}},
\code{\link{getDefaultGeomAttributes}()},
\code{\link{getDefaultOptions}()},
\code{\link{resetDefaultShapeDiscrete}()},
\code{\link{resetDefaults}()},
\code{\link{setDefaultColorMapDistinct}()},
\code{\link{setDefaultShapeDiscrete}()},
\code{\link{setDefaultTheme}()},
\code{\link{setDefaults}()}
}
\concept{setDefault functions}
