% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-defaults.R
\name{resetDefaults}
\alias{resetDefaults}
\title{restore to previously stored settings}
\usage{
resetDefaults(oldDefaults)
}
\arguments{
\item{oldDefaults}{\code{list}with previously stored settings}
}
\description{
restore to previously stored settings
}
\seealso{
Other setDefault functions: 
\code{\link{ColorMaps}},
\code{\link{getDefaultGeomAttributes}()},
\code{\link{getDefaultOptions}()},
\code{\link{resetDefaultColorMapDistinct}()},
\code{\link{resetDefaultShapeDiscrete}()},
\code{\link{setDefaultColorMapDistinct}()},
\code{\link{setDefaultShapeDiscrete}()},
\code{\link{setDefaultTheme}()},
\code{\link{setDefaults}()}
}
\concept{setDefault functions}
