% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-defaults.R
\name{resetDefaultShapeDiscrete}
\alias{resetDefaultShapeDiscrete}
\title{resets the scale for discrete shapes to ggplot default}
\usage{
resetDefaultShapeDiscrete(oldShapeValues = NULL)
}
\arguments{
\item{oldShapeValues}{shape values to set}
}
\description{
resets the scale for discrete shapes to ggplot default
}
\seealso{
Other setDefault functions: 
\code{\link{Shapes}},
\code{\link{colorMaps}},
\code{\link{geomPointUnicode}()},
\code{\link{getDefaultGeomAttributes}()},
\code{\link{getDefaultOptions}()},
\code{\link{getOspsuite.plots.option}()},
\code{\link{resetDefaultColorMapDistinct}()},
\code{\link{resetDefaultTheme}()},
\code{\link{resetDefaults}()},
\code{\link{setDefaultColorMapDistinct}()},
\code{\link{setDefaultShapeDiscrete}()},
\code{\link{setDefaultTheme}()},
\code{\link{setDefaults}()},
\code{\link{setOspsuite.plots.option}()}
}
\concept{setDefault functions}
