% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-defaults.R
\name{resetDefaultshapeDiscrete}
\alias{resetDefaultshapeDiscrete}
\title{resets the scale for discrete shapes to ggplot default}
\usage{
resetDefaultshapeDiscrete()
}
\description{
resets the scale for discrete shapes to ggplot default
}
\seealso{
Other setDefaultsOspsuite.plots: 
\code{\link{ColorMaps}},
\code{\link{getDefaultGeomAttributes}()},
\code{\link{getDefaultOptions}()},
\code{\link{resetDefaultColorMapDistinct}()},
\code{\link{resetDefaults}()},
\code{\link{setDefaultColorMapDistinct}()},
\code{\link{setDefaultTheme}()},
\code{\link{setDefaultshapeDiscrete}()},
\code{\link{setDefaults}()}
}
\concept{setDefaultsOspsuite.plots}
