% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-defaults.R
\name{resetDefaultShapeDiscrete}
\alias{resetDefaultShapeDiscrete}
\title{resets the scale for discrete shapes to ggplot default}
\usage{
resetDefaultShapeDiscrete()
}
\description{
resets the scale for discrete shapes to ggplot default
}
\seealso{
Other setDefault functions: 
\code{\link{ColorMaps}},
\code{\link{Shapes}},
\code{\link{geomPointUnicode}()},
\code{\link{getDefaultGeomAttributes}()},
\code{\link{getDefaultOptions}()},
\code{\link{resetDefaultColorMapDistinct}()},
\code{\link{resetDefaults}()},
\code{\link{setDefaultColorMapDistinct}()},
\code{\link{setDefaultShapeDiscrete}()},
\code{\link{setDefaultTheme}()},
\code{\link{setDefaults}()}
}
\concept{setDefault functions}
