% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-defaults.R
\name{resetDefaultTheme}
\alias{resetDefaultTheme}
\title{reset the default theme}
\usage{
resetDefaultTheme(oldTheme)
}
\arguments{
\item{oldTheme}{theme to set}
}
\description{
wrapper for \code{ggplot2::theme_set(oldTheme)}
}
\seealso{
Other setDefault functions: 
\code{\link{Shapes}},
\code{\link{colorMaps}},
\code{\link{geomPointUnicode}()},
\code{\link{getDefaultGeomAttributes}()},
\code{\link{getDefaultOptions}()},
\code{\link{getOspsuite.plots.option}()},
\code{\link{resetDefaultColorMapDistinct}()},
\code{\link{resetDefaultShapeDiscrete}()},
\code{\link{resetDefaults}()},
\code{\link{setDefaultColorMapDistinct}()},
\code{\link{setDefaultShapeDiscrete}()},
\code{\link{setDefaultTheme}()},
\code{\link{setDefaults}()},
\code{\link{setOspsuite.plots.option}()}
}
\concept{setDefault functions}
