% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-defaults.R
\name{setDefaultColorMapDistinct}
\alias{setDefaultColorMapDistinct}
\title{set the default colormap for discrete colors}
\usage{
setDefaultColorMapDistinct(colorMapList = NULL)
}
\arguments{
\item{colorMapList}{list of colorMaps to be set}
}
\value{
list with Colormaps previously set
}
\description{
set the default colormap for discrete colors
}
\seealso{
Other setDefault functions: 
\code{\link{Shapes}},
\code{\link{colorMaps}},
\code{\link{geomPointUnicode}()},
\code{\link{getDefaultGeomAttributes}()},
\code{\link{getDefaultOptions}()},
\code{\link{resetDefaultColorMapDistinct}()},
\code{\link{resetDefaultShapeDiscrete}()},
\code{\link{resetDefaultTheme}()},
\code{\link{resetDefaults}()},
\code{\link{setDefaultShapeDiscrete}()},
\code{\link{setDefaultTheme}()},
\code{\link{setDefaults}()}
}
\concept{setDefault functions}
