% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-defaults.R
\name{setDefaultColorMapDistinct}
\alias{setDefaultColorMapDistinct}
\title{set the default colormap for discrete colors}
\usage{
setDefaultColorMapDistinct(ColorMapList = NULL)
}
\arguments{
\item{ColorMapList}{list of ColorMaps to be set}
}
\value{
list with COlormaps previously set
}
\description{
set the default colormap for discrete colors
}
\seealso{
Other setDefault functions: 
\code{\link{ColorMaps}},
\code{\link{getDefaultGeomAttributes}()},
\code{\link{getDefaultOptions}()},
\code{\link{resetDefaultColorMapDistinct}()},
\code{\link{resetDefaultShapeDiscrete}()},
\code{\link{resetDefaults}()},
\code{\link{setDefaultShapeDiscrete}()},
\code{\link{setDefaultTheme}()},
\code{\link{setDefaults}()}
}
\concept{setDefault functions}
