% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-defaults.R
\name{setDefaultColorMapDistinct}
\alias{setDefaultColorMapDistinct}
\title{set the default colormap for discret colors}
\usage{
setDefaultColorMapDistinct(ColorMapList = NULL)
}
\arguments{
\item{ColorMapList}{list of ColorMaps to be set}
}
\value{
list wih COlormpas previously set
}
\description{
set the default colormap for discret colors
}
\seealso{
Other setDefaultsOspsuite.plots: 
\code{\link{ColorMaps}},
\code{\link{getDefaultGeomAttributes}()},
\code{\link{getDefaultOptions}()},
\code{\link{resetDefaultColorMapDistinct}()},
\code{\link{resetDefaultshapeDiscrete}()},
\code{\link{resetDefaults}()},
\code{\link{setDefaultTheme}()},
\code{\link{setDefaultshapeDiscrete}()},
\code{\link{setDefaults}()}
}
\concept{setDefaultsOspsuite.plots}
