% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-defaults.R
\name{setDefaults}
\alias{setDefaults}
\title{sets the defaults for the OSPSuite.plots package}
\usage{
setDefaults(defaultOptions = list(), colorMapList = NULL)
}
\arguments{
\item{defaultOptions}{list of options}

\item{colorMapList}{list of color maps}
}
\value{
list of old settings which can be used to reset defaults with \code{resetDefaults()}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

Mutates global \code{ggplot2} state (theme, geom defaults and discrete color
options) for the whole session.

\code{ospsuite.plots} plot functions now apply the full OSPSuite styling per plot,
so this is no longer needed for them. To style individual (non-\code{ospsuite.plots})
plots, compose the per-plot constructors instead:
\code{plot + themeOspsuite() + scale_colour_osp() + scale_fill_osp()}.

for detailed information see
\code{vignette("ospsuite.plots", package = "ospsuite.plots")}
}
\seealso{
Other setDefault functions:
\code{\link{colorMaps}},
\code{\link[=getDefaultGeomAttributes]{getDefaultGeomAttributes()}},
\code{\link[=getDefaultOptions]{getDefaultOptions()}},
\code{\link[=getOspsuite.plots.option]{getOspsuite.plots.option()}},
\code{\link[=resetDefaultColorMapDistinct]{resetDefaultColorMapDistinct()}},
\code{\link[=resetDefaultTheme]{resetDefaultTheme()}},
\code{\link[=resetDefaults]{resetDefaults()}},
\code{\link[=setDefaultColorMapDistinct]{setDefaultColorMapDistinct()}},
\code{\link[=setDefaultTheme]{setDefaultTheme()}},
\code{\link[=setOspsuite.plots.option]{setOspsuite.plots.option()}},
\code{\link[=themeOspsuite]{themeOspsuite()}}
}
\concept{setDefault functions}
