% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-defaults.R
\name{setDefaults}
\alias{setDefaults}
\title{sets the defaults for the OSPSuite.plots package}
\usage{
setDefaults(
  defaultOptions = list(),
  colorMapList = NULL,
  shapeValues = NULL,
  pointAsUnicode = FALSE
)
}
\arguments{
\item{defaultOptions}{list of options}

\item{colorMapList}{list of color maps}

\item{shapeValues}{list of Shapes}

\item{pointAsUnicode}{A \code{flag} to switch between mode for geom_point, if TRUE points will be plotted as unicode labels}
}
\value{
list of old settings which can be used to reset defaults with \code{resetDefaults()}
}
\description{
should be started at the beginning at each workflow
}
\details{
for detailed information see
\code{vignette("ospsuite.plots", package = "ospsuite.plots")}
}
\seealso{
Other setDefault functions: 
\code{\link{Shapes}},
\code{\link{colorMaps}},
\code{\link{geomPointUnicode}()},
\code{\link{getDefaultGeomAttributes}()},
\code{\link{getDefaultOptions}()},
\code{\link{getOspsuite.plots.option}()},
\code{\link{resetDefaultColorMapDistinct}()},
\code{\link{resetDefaultShapeDiscrete}()},
\code{\link{resetDefaultTheme}()},
\code{\link{resetDefaults}()},
\code{\link{setDefaultColorMapDistinct}()},
\code{\link{setDefaultShapeDiscrete}()},
\code{\link{setDefaultTheme}()},
\code{\link{setOspsuite.plots.option}()}
}
\concept{setDefault functions}
