% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-defaults.R
\name{setDefaults}
\alias{setDefaults}
\title{sets the defaults for Ospsuite.plot}
\usage{
setDefaults(
  ColorMapList = NULL,
  shapeValues = NULL,
  defaultOptions = getDefaultOptions()
)
}
\arguments{
\item{ColorMapList}{list of color maps}

\item{shapeValues}{list of Shapes}

\item{defaultOptions}{list of options}
}
\value{
list of old settings which can be used to reset defaults with restggOPSDefaults
}
\description{
should be started at the beginning at each workflow
}
\details{
for detailde information see
\code{vignette("ospsuite.plots", package = "ospsuite.plots")}
}
\seealso{
Other setDefaultsOspsuite.plots: 
\code{\link{ColorMaps}},
\code{\link{getDefaultGeomAttributes}()},
\code{\link{getDefaultOptions}()},
\code{\link{resetDefaultColorMapDistinct}()},
\code{\link{resetDefaultshapeDiscrete}()},
\code{\link{resetDefaults}()},
\code{\link{setDefaultColorMapDistinct}()},
\code{\link{setDefaultTheme}()},
\code{\link{setDefaultshapeDiscrete}()}
}
\concept{setDefaultsOspsuite.plots}
