% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-defaults.R
\name{setDefaultShapeDiscrete}
\alias{setDefaultShapeDiscrete}
\title{set the default shapes scale for discrete values}
\usage{
setDefaultShapeDiscrete(shapeValues = NULL)
}
\arguments{
\item{shapeValues}{vector of shapevalues}
}
\description{
set the default shapes scale for discrete values
}
\seealso{
Other setDefaultsOspsuite.plots: 
\code{\link{ColorMaps}},
\code{\link{getDefaultGeomAttributes}()},
\code{\link{getDefaultOptions}()},
\code{\link{resetDefaultColorMapDistinct}()},
\code{\link{resetDefaultShapeDiscrete}()},
\code{\link{resetDefaults}()},
\code{\link{setDefaultColorMapDistinct}()},
\code{\link{setDefaultTheme}()},
\code{\link{setDefaults}()}
}
\concept{setDefaultsOspsuite.plots}
