% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-defaults.R
\name{setDefaultShapeDiscrete}
\alias{setDefaultShapeDiscrete}
\title{set the default shapes}
\usage{
setDefaultShapeDiscrete(shapeValues = NULL)
}
\arguments{
\item{shapeValues}{vector of shapevalues}
}
\value{
vector with shapeValues saved in option ospsuite.plots.shapeValues before function call
}
\description{
The scales are set to the option  'ospsuite.plots.shapeValues', which is the used to
set the discrete scale of shapes for all {ospsuite.plots} function
for customized functions add scale_shape_manual(values = getOption('ospsuite.plots.shapeValues'))
}
\seealso{
Other setDefault functions: 
\code{\link{ColorMaps}},
\code{\link{Shapes}},
\code{\link{geomPointUnicode}()},
\code{\link{getDefaultGeomAttributes}()},
\code{\link{getDefaultOptions}()},
\code{\link{resetDefaultColorMapDistinct}()},
\code{\link{resetDefaultShapeDiscrete}()},
\code{\link{resetDefaultTheme}()},
\code{\link{resetDefaults}()},
\code{\link{setDefaultColorMapDistinct}()},
\code{\link{setDefaultTheme}()},
\code{\link{setDefaults}()}
}
\concept{setDefault functions}
