% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-defaults.R
\name{setDefaultShapeDiscrete}
\alias{setDefaultShapeDiscrete}
\title{set the default shapes}
\usage{
setDefaultShapeDiscrete(shapeValues = NULL)
}
\arguments{
\item{shapeValues}{vector of shape values}
}
\value{
vector with \code{shapeValues} saved in option \code{ospsuite.plots.shapeValues} before function call
}
\description{
The scales are set to the option  \code{ospsuite.plots.shapeValues}, which is the used to
set the discrete scale of shapes for all \code{ospsuite.plots} function
for customized functions add scale_shape_manual(\code{values = getOspsuite.plots.option(optionKey = OptionKeys$shapeValues)})
}
\seealso{
Other setDefault functions: 
\code{\link{Shapes}},
\code{\link{colorMaps}},
\code{\link{geomPointUnicode}()},
\code{\link{getDefaultGeomAttributes}()},
\code{\link{getDefaultOptions}()},
\code{\link{getOspsuite.plots.option}()},
\code{\link{resetDefaultColorMapDistinct}()},
\code{\link{resetDefaultShapeDiscrete}()},
\code{\link{resetDefaultTheme}()},
\code{\link{resetDefaults}()},
\code{\link{setDefaultColorMapDistinct}()},
\code{\link{setDefaultTheme}()},
\code{\link{setDefaults}()},
\code{\link{setOspsuite.plots.option}()}
}
\concept{setDefault functions}
