% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-defaults.R
\name{setDefaultTheme}
\alias{setDefaultTheme}
\title{set the default theme}
\usage{
setDefaultTheme()
}
\value{
invisibly return the previous theme so you can easily save it, then later restore it.
}
\description{
set properties of the default theme
}
\seealso{
Other setDefault functions: 
\code{\link{Shapes}},
\code{\link{colorMaps}},
\code{\link{geomPointUnicode}()},
\code{\link{getDefaultGeomAttributes}()},
\code{\link{getDefaultOptions}()},
\code{\link{getOspsuite.plots.option}()},
\code{\link{resetDefaultColorMapDistinct}()},
\code{\link{resetDefaultShapeDiscrete}()},
\code{\link{resetDefaultTheme}()},
\code{\link{resetDefaults}()},
\code{\link{setDefaultColorMapDistinct}()},
\code{\link{setDefaultShapeDiscrete}()},
\code{\link{setDefaults}()},
\code{\link{setOspsuite.plots.option}()}
}
\concept{setDefault functions}
