% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-defaults.R
\name{setOspsuite.plots.option}
\alias{setOspsuite.plots.option}
\title{Set OSPSuite plots option with a given key and value.}
\usage{
setOspsuite.plots.option(optionKey, value)
}
\arguments{
\item{optionKey}{The key for the option.}

\item{value}{The value for the option.}
}
\description{
Set OSPSuite plots option with a given key and value.
}
\examples{
\dontrun{
setOspsuite.plots.option(optionKey = OptionKeys$watermark_enabled, value = TRUE)
}
}
\seealso{
Other setDefault functions: 
\code{\link{Shapes}},
\code{\link{colorMaps}},
\code{\link{geomPointUnicode}()},
\code{\link{getDefaultGeomAttributes}()},
\code{\link{getDefaultOptions}()},
\code{\link{getOspsuite.plots.option}()},
\code{\link{resetDefaultColorMapDistinct}()},
\code{\link{resetDefaultShapeDiscrete}()},
\code{\link{resetDefaultTheme}()},
\code{\link{resetDefaults}()},
\code{\link{setDefaultColorMapDistinct}()},
\code{\link{setDefaultShapeDiscrete}()},
\code{\link{setDefaultTheme}()},
\code{\link{setDefaults}()}
}
\concept{setDefault functions}
