% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-defaults.R
\name{setOspsuite.plots.option}
\alias{setOspsuite.plots.option}
\title{Set OSPSuite plots option with a given key and value.}
\usage{
setOspsuite.plots.option(optionKey, value)
}
\arguments{
\item{optionKey}{The key for the option.}

\item{value}{The value for the option.}
}
\description{
Set OSPSuite plots option with a given key and value.
}
\examples{
\dontrun{
setOspsuite.plots.option(optionKey = OptionKeys$watermark_enabled, value = TRUE)
}
}
