% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_shapes.R
\name{Shapes}
\alias{Shapes}
\title{Shapes}
\usage{
Shapes
}
\description{
Named list of OSP shape names for backward compatibility.
Use \code{Shapes$circle} to get the shape name "circle".
}
\seealso{
Other setDefault functions:
\code{\link{colorMaps}},
\code{\link[=geom_point_osp]{geom_point_osp()}},
\code{\link[=getDefaultGeomAttributes]{getDefaultGeomAttributes()}},
\code{\link[=getDefaultOptions]{getDefaultOptions()}},
\code{\link[=getOspsuite.plots.option]{getOspsuite.plots.option()}},
\code{\link[=resetDefaultColorMapDistinct]{resetDefaultColorMapDistinct()}},
\code{\link[=resetDefaultTheme]{resetDefaultTheme()}},
\code{\link[=resetDefaults]{resetDefaults()}},
\code{\link[=scale_shape_osp]{scale_shape_osp()}},
\code{\link[=scale_shape_osp_identity]{scale_shape_osp_identity()}},
\code{\link[=scale_shape_osp_manual]{scale_shape_osp_manual()}},
\code{\link[=setDefaultColorMapDistinct]{setDefaultColorMapDistinct()}},
\code{\link[=setDefaultTheme]{setDefaultTheme()}},
\code{\link[=setDefaults]{setDefaults()}},
\code{\link[=setOspsuite.plots.option]{setOspsuite.plots.option()}},
\code{\link[=stat_qq_osp]{stat_qq_osp()}}
}
\concept{setDefault functions}
