% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_shapes.R
\docType{data}
\name{Shapes}
\alias{Shapes}
\title{Shapes}
\format{
An object of class \code{list} of length 22.
}
\usage{
Shapes
}
\description{
Named list of OSP shape names for backward compatibility.
Use \code{Shapes$circle} to get the shape name "circle".
}
\seealso{
Other setDefault functions: 
\code{\link{colorMaps}},
\code{\link{geom_point_osp}()},
\code{\link{getDefaultGeomAttributes}()},
\code{\link{getDefaultOptions}()},
\code{\link{getOspsuite.plots.option}()},
\code{\link{resetDefaultColorMapDistinct}()},
\code{\link{resetDefaultShapeDiscrete}()},
\code{\link{resetDefaultTheme}()},
\code{\link{resetDefaults}()},
\code{\link{scale_shape_osp}()},
\code{\link{scale_shape_osp_identity}()},
\code{\link{scale_shape_osp_manual}()},
\code{\link{setDefaultColorMapDistinct}()},
\code{\link{setDefaultShapeDiscrete}()},
\code{\link{setDefaultTheme}()},
\code{\link{setDefaults}()},
\code{\link{setOspsuite.plots.option}()},
\code{\link{stat_qq_osp}()}
}
\concept{setDefault functions}
\keyword{datasets}
