% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_shapes.R
\docType{data}
\name{Shapes}
\alias{Shapes}
\title{Shapes}
\format{
An object of class \code{list} of length 40.
}
\usage{
Shapes
}
\description{
List of some \code{ggplot2} shapes.
The shapes from this list are unicode characters
corresponding to their appropriate shapes.
}
\seealso{
Other setDefault functions: 
\code{\link{colorMaps}},
\code{\link{geomPointUnicode}()},
\code{\link{getDefaultGeomAttributes}()},
\code{\link{getDefaultOptions}()},
\code{\link{getOspsuite.plots.option}()},
\code{\link{resetDefaultColorMapDistinct}()},
\code{\link{resetDefaultShapeDiscrete}()},
\code{\link{resetDefaultTheme}()},
\code{\link{resetDefaults}()},
\code{\link{setDefaultColorMapDistinct}()},
\code{\link{setDefaultShapeDiscrete}()},
\code{\link{setDefaultTheme}()},
\code{\link{setDefaults}()},
\code{\link{setOspsuite.plots.option}()}
}
\concept{setDefault functions}
\keyword{datasets}
