% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{updateScaleArgumentsForTimeUnit}
\alias{updateScaleArgumentsForTimeUnit}
\title{adjust arguments for scale if dimension of scale is time}
\usage{
updateScaleArgumentsForTimeUnit(
  scale.args,
  metaData,
  mapping,
  scaleDirection = "x"
)
}
\arguments{
\item{scale.args}{list of arguments for scale to be updated}

\item{metaData}{A named list of information about the \code{data} such as the \code{dimension} and \code{unit} of its variables.}

\item{mapping}{Default list of aesthetic mappings to use for plot}

\item{scaleDirection}{either 'x' or 'y'}
}
\value{
update list of arguments for scale
}
\description{
use metadata and mapping to determine dimension of axis
use fixed width for breaks depending on unit:
}
\details{
\itemize{
\item s: width = 15,
\item min: width = 15,
\item h: width = 6,
\item day(s): width = 7
\item week(s): width = 4
\item month(s): width = 6
}

use width only if 2 * width < range of time values
use multiples of width if 20 * width < range of time values
}
\examples{
newScaleArguments <- updateScaleArgumentsForTimeUnit(
  scale.args = list(),
  metaData = list(time = list(dimension = "Time", unit = "h")),
  mapping = aes(x = time)
)

}
