% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_export.R
\name{validateFilename}
\alias{validateFilename}
\title{Replace special letters in file names}
\usage{
validateFilename(filename, device)
}
\arguments{
\item{filename}{Name of the file to validate.}

\item{device}{A character with the device to use}
}
\value{
File name without special letters.
}
\description{
Replace special letters in file names
}
