% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_export.R
\name{validateInputsExportPlot}
\alias{validateInputsExportPlot}
\title{Validates the inputs for the plot export}
\usage{
validateInputsExportPlot(plotObject, filepath, filename, width, height)
}
\arguments{
\item{plotObject}{ggplot object}

\item{filepath}{name of export directory}

\item{filename}{name of file to save.
extension is added or switched to default device
\code{getOspsuite.plots.option(optionKey = OptionKeys$export.device)}}

\item{width}{\code{double} width of exported file, defaults is \code{getOption('ospsuite.plots.export.width',default = 16)}
unit \code{getOption('ospsuite.plots.export.units',  default = "cm")}}

\item{height}{\code{double} height of exported plot, if NULL the aspect ratio of plot is used,
or if no aspect ratio is defined, a square plot is exported}
}
\description{
Validates the inputs for the plot export
}
